Re-validate with eval removal + code shrinking
Changes from previous run:
- Removed mid-training and diagnostic eval_val calls (reclaims ~10-15s for training)
- Shrunk train_gpt.py via AST pruning + pyminify + LZMA (125KB -> 24KB, frees ~99KB code budget)
- Human-readable source preserved as train_gpt_human.py

3-seed mean val_bpb (SW s64): 1.1091
Max artifact bytes: 15993904
