+++ records/track_10min_16mb/2026-03-28_TurboMuon_EngramLite_ParamBanking/README.md
@@ -1,15 +1,15 @@
 # Turbo-Muon + EngramLite + Parameter Banking + GPTQ Mixed-Precision
 
-**val_bpb: 1.1086** (3-seed mean, std 0.0006) | **~15.3 MB** | 8xH100 SXM
+**val_bpb: 1.1091** (3-seed mean, std 0.0005) | **~15.3 MB** | 8xH100 SXM
 
 ## Results (8xH100 80GB SXM)
 
 | Seed | step_avg | steps | val_bpb (SW) | val_bpb (full) | Artifact bytes |
 |------|----------|-------|-------------|----------------|----------------|
-| 42   | 93.68ms  | 6257  | 1.1079      | 1.1317         | 15,986,468     |
-| 1337 | 93.12ms  | 6294  | 1.1089      | 1.1326         | 15,996,753     |
-| 2025 | 93.08ms  | 6297  | 1.1090      | 1.1327         | 15,997,089     |
-| **Mean** | **93.29ms** | **6283** | **1.1086** | **1.1324** | |
+| 42   | 93.26ms  | 6284  | 1.1086      | 1.1324         | 15,992,528     |
+| 1337 | 93.11ms  | 6295  | 1.1090      | 1.1328         | 15,993,413     |
+| 2025 | 93.11ms  | 6294  | 1.1096      | 1.1335         | 15,993,904     |
+| **Mean** | **93.16ms** | **6291** | **1.1091** | **1.1329** | |
 
 ## Summary
 
@@ -75,6 +75,16 @@ Compression pipeline with Hessian collection performed within the 600s training 
 4. **Late QAT (soft-round)** — Quantization-aware training activated when LR scale drops below 15%, with soft-round sigmoid alpha ramping 1→16 over the QAT phase. Provides real gradient signal through quantization grid points.
 5. **Selective pruning** — Post-GPTQ, values with `|q| ≤ 2` ranked by reconstruction error impact. Binary search with fast (zlib-1) / real (brotli-11) calibration finds the minimal prune count to fit 16MB.
 6. **Brotli + byte-shuffle compression** — Byte-shuffle preprocessing reorders tensor bytes by significance position before brotli compression (quality=11) for optimal entropy coding.
+
+### Code Shrinking
+
+The submission `train_gpt.py` is a compressed self-extracting wrapper generated by
+`Shrink/shrink.py`. The pipeline: AST dead-code removal → pyminify (strip comments,
+whitespace, type hints, rename identifiers) → LZMA + base85 self-extracting `exec()`.
+
+- **Human-readable source**: `train_gpt_human.py` (123 KB)
+- **Shrunk submission**: `train_gpt.py` (24 KB)
+- **Code budget freed**: ~99 KB → more artifact space for model weights → less pruning
 
 ## Architecture
 
@@ -153,10 +163,3 @@ LOAD_SNAPSHOT=snapshot_post_hessian.pt torchrun --standalone --nproc_per_node=1 
 - **Polar Express coefficients**: Amsel et al. (arXiv:2505.16932)
 - **GPTQ approach**: [PR #634](https://github.com/openai/parameter-golf/pull/634) Hessian-aware quantization
 
-## Validation Checklist
-
-- [x] Run 3 seeds on 8xH100 (42, 1337, 2025) and collect logs
-- [x] Fill in results table with actual val_bpb, step times, artifact sizes
-- [x] Update submission.json with actual val_bpb and bytes_total
-- [ ] Verify 3-seed mean beats SOTA by >= 0.005 nats
-- [x] Attach log files as `train_seed42.log`, `train_seed1337.log`, `train_seed2025.log`
