+++ records/track_10min_16mb/2026-03-28_TurboMuon_EngramLite_ParamBanking/submission.json
@@ -1,7 +1,7 @@
 {
   "name": "Turbo-Muon + EngramLite + Parameter Banking + GPTQ Mixed-Precision",
-  "val_bpb": "1.1086",
-  "bytes_total": "15997089",
+  "val_bpb": "1.1091",
+  "bytes_total": "15993904",
   "blurb": "11L/512d GPT with Turbo-Muon (AOL+Polar Express+row_col norm), EngramLite bigram+trigram hash embeddings, U-Net skip connections, ValueEmbedding, SmearGate, SWA+EMA, GPTQ mixed-precision int6/int7 with Hessian sensitivity, brotli+byte-shuffle compression. Built on PR #609.",
   "author": "mikeapedia",
   "github_id": "mikeapedia",
