+++ records/track_10min_16mb/2026-03-28_TurboMuon_EngramLite_ParamBanking/train_gpt.py
@@ -1,2565 +1,237 @@
-from __future__ import annotations
-import copy
-import glob
-import io
-import lzma
-import math
-import os
-import random
-import subprocess
-import sys
-import time
-import uuid
-import zlib
-from pathlib import Path
-
-import numpy as np
-import sentencepiece as spm
-import torch
-import torch.distributed as dist
-import torch.nn.functional as F
-from torch import Tensor, nn
-from torch.nn.parallel import DistributedDataParallel as DDP
-
-# Compression preference: brotli > lzma > zstd > zlib
-# brotli q=11 typically beats lzma preset=9 by 1-5% on quantized weights
-try:
-    import brotli as _brotli_probe  # noqa: F401
-    _COMPRESSOR = "brotli"
-except ImportError:
-    _COMPRESSOR = "lzma"
-
-# Byte-shuffle preprocessing: reorder bytes by stride position before compression.
-# For multi-byte values (float16 scales), grouping same-position bytes
-# together creates runs of similar values -> better entropy coding. Lossless & fast (<1s).
-_BYTE_SHUFFLE = True
-_BYTE_SHUFFLE_STRIDE = 2  # 2 = optimal for float16-heavy data
-
-_BSHF_MAGIC = b"BSHF"  # 4-byte magic for byte-shuffled data
-
-
-def _byte_shuffle(data: bytes, stride: int = 2) -> bytes:
-    """Transpose byte stream by stride using numpy vectorized indexing.
-
-    Groups byte[i % stride] positions together. For stride=2 on float16 data,
-    this puts all high bytes together and all low bytes together, dramatically
-    improving compression. Prepend 4-byte magic + 1-byte stride header.
-    """
-    if stride <= 1 or len(data) < stride:
-        return data  # no-op
-    src = np.frombuffer(data, dtype=np.uint8)
-    n = len(src)
-    out = np.empty(n, dtype=np.uint8)
-    dest_off = 0
-    for pos in range(stride):
-        chunk = src[pos::stride]  # every stride-th byte starting at pos
-        out[dest_off:dest_off + len(chunk)] = chunk
-        dest_off += len(chunk)
-    return _BSHF_MAGIC + bytes([stride]) + out.tobytes()
-
-
-def _byte_unshuffle(data: bytes) -> bytes:
-    """Inverse of _byte_shuffle using numpy. Auto-detects BSHF magic header."""
-    if len(data) < 5 or data[:4] != _BSHF_MAGIC:
-        return data  # no shuffle header -> old format, return as-is
-    stride = data[4]
-    if stride < 2:
-        return data[5:]  # invalid stride, just strip header
-    payload = np.frombuffer(data, dtype=np.uint8, offset=5)
-    n = len(payload)
-    out = np.empty(n, dtype=np.uint8)
-    src_off = 0
-    for pos in range(stride):
-        chunk_len = n // stride + (1 if pos < n % stride else 0)
-        out[pos::stride][:chunk_len] = payload[src_off:src_off + chunk_len]
-        src_off += chunk_len
-    return out.tobytes()
-
-
-class Hyperparameters:
-    data_path = os.environ.get("DATA_PATH", "./data/datasets/fineweb10B_sp1024")
-    train_files = os.path.join(data_path, "fineweb_train_*.bin")
-    val_files = os.path.join(data_path, "fineweb_val_*.bin")
-    tokenizer_path = os.environ.get("TOKENIZER_PATH", "./data/tokenizers/fineweb_1024_bpe.model")
-    run_id = os.environ.get("RUN_ID", str(uuid.uuid4()))
-    seed = int(os.environ.get("SEED", 1337))
-    val_batch_size = int(os.environ.get("VAL_BATCH_SIZE", 524_288))
-
-    train_log_every = int(os.environ.get("TRAIN_LOG_EVERY", 500))
-    iterations = int(os.environ.get("ITERATIONS", 20000))
-    warmdown_iters = int(os.environ.get("WARMDOWN_ITERS", 3500))
-    lr_floor = float(os.environ.get("LR_FLOOR", 0.05))  # Minimum LR as fraction of peak (prevents sharp quant-sensitive minima)
-    warmup_steps = int(os.environ.get("WARMUP_STEPS", 20))
-    train_batch_tokens = int(os.environ.get("TRAIN_BATCH_TOKENS", 786_432))
-    train_seq_len = int(os.environ.get("TRAIN_SEQ_LEN", 2048))
-    eval_seq_len = int(os.environ.get("EVAL_SEQ_LEN", 2048))
-    max_wallclock_seconds = float(os.environ.get("MAX_WALLCLOCK_SECONDS", 600.0))
-    qk_gain_init = float(os.environ.get("QK_GAIN_INIT", 1.5))
-    vocab_size = int(os.environ.get("VOCAB_SIZE", 1024))
-    num_layers = int(os.environ.get("NUM_LAYERS", 11))
-    num_kv_heads = int(os.environ.get("NUM_KV_HEADS", 4))
-    model_dim = int(os.environ.get("MODEL_DIM", 512))
-    num_heads = int(os.environ.get("NUM_HEADS", 8))
-    mlp_mult = float(os.environ.get("MLP_MULT", 3.5))
-    tie_embeddings = bool(int(os.environ.get("TIE_EMBEDDINGS", "1")))
-    rope_base = float(os.environ.get("ROPE_BASE", 10000.0))
-    logit_softcap = float(os.environ.get("LOGIT_SOFTCAP", 30.0))
-    embed_lr = float(os.environ.get("EMBED_LR", 0.6))
-    head_lr = float(os.environ.get("HEAD_LR", 0.008))
-    tied_embed_lr = float(os.environ.get("TIED_EMBED_LR", 0.035))
-    tied_embed_init_std = float(os.environ.get("TIED_EMBED_INIT_STD", 0.005))
-    matrix_lr = float(os.environ.get("MATRIX_LR", 0.025))
-    scalar_lr = float(os.environ.get("SCALAR_LR", 0.025))
-    muon_momentum = float(os.environ.get("MUON_MOMENTUM", 0.99))
-    muon_backend_steps = int(os.environ.get("MUON_BACKEND_STEPS", 4))
-    muon_momentum_warmup_start = float(os.environ.get("MUON_MOMENTUM_WARMUP_START", 0.92))
-    muon_momentum_warmup_steps = int(os.environ.get("MUON_MOMENTUM_WARMUP_STEPS", 1500))
-    beta1 = float(os.environ.get("BETA1", 0.9))
-    beta2 = float(os.environ.get("BETA2", 0.95))
-    adam_eps = float(os.environ.get("ADAM_EPS", 1e-8))
-    grad_clip_norm = float(os.environ.get("GRAD_CLIP_NORM", 0.3))
-    eval_stride = int(os.environ.get("EVAL_STRIDE", 64))
-    swa_enabled = bool(int(os.environ.get("SWA_ENABLED", "1")))
-    swa_every = int(os.environ.get("SWA_EVERY", 50))
-    swa_threshold = float(os.environ.get("SWA_THRESHOLD", 0.2))
-    muon_wd = float(os.environ.get("MUON_WD", 0.04))
-    adam_wd = float(os.environ.get("ADAM_WD", 0.04))
-    xsa_last_n = int(os.environ.get("XSA_LAST_N", 11))
-    rope_dims = int(os.environ.get("ROPE_DIMS", 16))
-    ln_scale = bool(int(os.environ.get("LN_SCALE", "1")))
-    ve_enabled = bool(int(os.environ.get("VE_ENABLED", "1")))
-    ve_dim = int(os.environ.get("VE_DIM", 128))
-    ve_layers = os.environ.get("VE_LAYERS", "9,10")
-    ema_enabled = bool(int(os.environ.get("EMA_ENABLED", "1")))
-    ema_decay = float(os.environ.get("EMA_DECAY", 0.997))
-    # New hyperparameters from our improvements
-    embed_beta1 = float(os.environ.get("EMBED_BETA1", 0.7))
-    head_beta1 = float(os.environ.get("HEAD_BETA1", 0.7))
-    muon_post_norm = os.environ.get("MUON_POST_NORM", "row_col")
-    qat_threshold = float(os.environ.get("QAT_THRESHOLD", 0.15))
-    qat_clip_pct = float(os.environ.get("QAT_CLIP_PCT", 0.9995))
-    late_qat = bool(int(os.environ.get("LATE_QAT", "1")))
-    mixed_precision = bool(int(os.environ.get("MIXED_PRECISION", "1")))
-    # mp_promote removed — dynamic allocation based on artifact size budget (see _allocate_bits_mixed)
-    target_bytes_limit = int(os.environ.get("TARGET_BYTES", 16_000_000))
-    # N-gram params
-    ngram_buckets = int(os.environ.get("NGRAM_BUCKETS", 8192))
-    ngram_heads = int(os.environ.get("NGRAM_HEADS", 2))
-    ngram_orders = int(os.environ.get("NGRAM_ORDERS", 2))
-    ngram_dim_per_head = int(os.environ.get("NGRAM_DIM_PER_HEAD", 32))
-    # GPTQ calibration
-    gptq_calib_batches = int(os.environ.get("GPTQ_CALIB_BATCHES", 64))
-    gptq_block_size = int(os.environ.get("GPTQ_BLOCK_SIZE", 128))
-    gptq_damp = float(os.environ.get("GPTQ_DAMP", 0.01))
-    gptq_reserve_ms = float(os.environ.get("GPTQ_RESERVE_MS", 14000.0))  # Reserve from training budget for GPTQ calibration (PR #634)
-    gptq_col_order = os.environ.get("GPTQ_COL_ORDER", "desc")  # "desc" (actorder) or "asc" (PR#753-style)
-    gptq_single_pass = bool(int(os.environ.get("GPTQ_SINGLE_PASS", "1")))  # Pre-compute scales, run GPTQ once
-    soft_round_qat = bool(int(os.environ.get("SOFT_ROUND_QAT", "1")))  # 1=soft-round, 0=STE
-    # Snapshot: save unbanked_sd + hessians after training, or load them to skip training
-    snapshot_post_hessian = bool(int(os.environ.get("SNAPSHOT_POST_HESSIAN", "0")))  # save snapshot and exit
-    load_snapshot = os.environ.get("LOAD_SNAPSHOT", "")  # path to snapshot file; skips training if set
-
-# --- Turbo-Muon Newton-Schulz orthogonalization ---
-
-# Polar Express optimal degree-5 coefficients (Amsel et al., arXiv:2505.16932).
-# With AOL preconditioning we skip iter 1 -- AOL already contracts the singular value
-# range via Gershgorin scaling, so we start from iteration 2.
-_POLAR_COEFFS_FULL = [
-    (8.28721201814563,  -23.595886519098837,  17.300387312530933),   # iter 1 (Frobenius init only)
-    (4.107059111542203,  -2.9478499167379106,  0.5448431082926601),  # iter 2
-    (3.9486908534822946, -2.908902115962949,   0.5518191394370137),  # iter 3
-    (3.3184196573706015, -2.488488024314874,   0.51004894012372),    # iter 4
-    (2.300652019954817,  -1.6689039845747493,  0.4188073119525673),  # iter 5
-    (1.891301407787398,  -1.2679958271945868,  0.37680408948524835), # iter 6
-    (1.875, -1.25, 0.375),  # iter 7+: converged fixed point
-]
-_AOL_POLAR_COEFFS = _POLAR_COEFFS_FULL[1:]  # skip iter 1 when using AOL
-
-
-def zeropower_via_newtonschulz5(G: Tensor, steps: int = 4, eps: float = 1e-7) -> Tensor:
-    """Turbo-Muon: Newton-Schulz with left-Gram AOL + Polar Express coefficients.
-
-    Supports both 2D (M, N) and batched 3D (B, M, N) input.
-    Uses left Gram (X@X.T) throughout -- always (m*m) where m <= n, giving
-    tighter Gershgorin bounds and up to 9x cheaper matmuls for non-square matrices.
-    """
-    X = G.bfloat16()
-    if X.ndim == 2:
-        # --- 2D path (single matrix) ---
-        transposed = X.size(0) > X.size(1)
-        if transposed:
-            X = X.T
-        A = X @ X.T
-        s = 1.0 / (A.abs().sum(dim=1).sqrt() + eps)
-        X = s.unsqueeze(1) * X
-        A = s.unsqueeze(0) * A * s.unsqueeze(1)
-        for i in range(steps):
-            a, b, c = _AOL_POLAR_COEFFS[min(i, len(_AOL_POLAR_COEFFS) - 1)]
-            if i > 0:
-                A = X @ X.T
-            B = b * A + c * A @ A
-            X = a * X + B @ X
-        return X.T if transposed else X
-    else:
-        # --- 3D batched path (B, M, N) ---
-        transposed = X.size(-2) > X.size(-1)
-        if transposed:
-            X = X.mT
-        A = X @ X.mT                                        # (B, m, m)
-        s = 1.0 / (A.abs().sum(dim=-1).sqrt() + eps)        # (B, m)
-        X = s.unsqueeze(-1) * X                              # (B, m, n)
-        A = s.unsqueeze(-2) * A * s.unsqueeze(-1)            # (B, m, m)
-        for i in range(steps):
-            a, b, c = _AOL_POLAR_COEFFS[min(i, len(_AOL_POLAR_COEFFS) - 1)]
-            if i > 0:
-                A = X @ X.mT
-            B = b * A + c * A @ A
-            X = a * X + B @ X
-        return X.mT if transposed else X
-
-
-def _post_ns_normalize(X: Tensor, mode: str) -> Tensor:
-    """Muon+ post-NS normalization: equalize per-neuron update magnitudes.
-
-    Modes: "none" (passthrough), "row", "col", "row_col".
-    Supports both 2D and batched 3D input.
-    Norms computed in float32 for numerical stability (X is typically bf16 from NS).
-    """
-    if mode == "none":
-        return X
-    if mode in ("row", "row_col"):
-        X = X / (X.float().norm(dim=-1, keepdim=True).to(X.dtype) + 1e-7)
-    if mode in ("col", "row_col"):
-        X = X / (X.float().norm(dim=-2, keepdim=True).to(X.dtype) + 1e-7)
-    return X
-
-
-# --- Parallel Muon optimizer ---
-
-class Muon(torch.optim.Optimizer):
-    """Parallel Muon: post-backward reduce-scatter -> local NS5 -> all-gather.
-
-    No DDP for bank params. After backward, this optimizer:
-    1. Launches async reduce-scatter for all banks (biggest first)
-    2. Returns control so Adam can step on small params while RS is in-flight
-    3. Waits for each RS, runs local NS5 on the shard, launches async all-gather
-    4. Each all-gather overlaps with next bank's NS5
-    """
-    def __init__(self, params, lr: float, momentum: float, backend_steps: int,
-                 nesterov: bool = True, weight_decay: float = 0.0,
-                 post_norm: str = "none"):
-        super().__init__(
-            params,
-            dict(lr=lr, momentum=momentum, backend_steps=backend_steps,
-                 nesterov=nesterov, weight_decay=weight_decay, post_norm=post_norm),
-        )
-        self._built = False
-
-    def _build(self):
-        self._distributed = dist.is_available() and dist.is_initialized()
-        self._world_size = dist.get_world_size() if self._distributed else 1
-        self._rank = dist.get_rank() if self._distributed else 0
-        ws = self._world_size
-
-        self._bank_meta = []
-        for group in self.param_groups:
-            for p in group["params"]:
-                B = p.shape[0]
-                padded_B = ((B + ws - 1) // ws) * ws
-                shard_B = padded_B // ws
-                tail = p.shape[1:]
-                dev = p.device
-                self._bank_meta.append({
-                    'p': p,
-                    'B': B,
-                    'padded_grad': torch.zeros(padded_B, *tail, device=dev, dtype=torch.bfloat16),
-                    'shard': torch.zeros(shard_B, *tail, device=dev, dtype=torch.bfloat16),
-                    'shard_mom': torch.zeros(shard_B, *tail, device=dev, dtype=torch.bfloat16),
-                    'full_update': torch.zeros(padded_B, *tail, device=dev, dtype=torch.bfloat16),
-                    'scale': max(1, p.shape[-2] / p.shape[-1]) ** 0.5,
-                })
-        # Sort by size descending -- launch biggest reduce-scatters first
-        self._bank_meta.sort(key=lambda m: -m['p'].numel())
-        self._built = True
-
-    def launch_reduce_scatters(self):
-        """Phase 1: launch async reduce-scatter for all banks. Call right after backward."""
-        if not self._built:
-            self._build()
-        if not self._distributed:
-            return
-        self._rs_futures = []
-        for m in self._bank_meta:
-            p = m['p']
-            if p.grad is None:
-                self._rs_futures.append(None)
-                continue
-            pg = m['padded_grad']
-            pg[:m['B']].copy_(p.grad.bfloat16())
-            if pg.shape[0] > m['B']:
-                pg[m['B']:].zero_()
-            fut = dist.reduce_scatter_tensor(m['shard'], pg, op=dist.ReduceOp.AVG, async_op=True)
-            self._rs_futures.append(fut)
-
-    @torch.no_grad()
-    def step(self, closure=None):
-        """Phase 3: wait for RS, local NS5, all-gather. Call AFTER Adam steps."""
-        loss = None
-        if closure is not None:
-            with torch.enable_grad():
-                loss = closure()
-
-        if not self._built:
-            self._build()
-
-        for group in self.param_groups:
-            lr = group["lr"]
-            momentum = group["momentum"]
-            backend_steps = group["backend_steps"]
-            nesterov = group["nesterov"]
-            wd = group.get("weight_decay", 0.0)
-            post_norm = group.get("post_norm", "none")
-
-            prev_ag_handle = None
-            prev_m = None
-
-            sharded = self._distributed and hasattr(self, '_rs_futures')
-
-            for i, m in enumerate(self._bank_meta):
-                p = m['p']
-                if p.grad is None:
-                    continue
-
-                if prev_ag_handle is not None:
-                    prev_ag_handle.wait()
-                    pp = prev_m['p']
-                    upd = prev_m['full_update'][:prev_m['B']]
-                    pp.add_(upd.to(dtype=pp.dtype), alpha=-lr * prev_m['scale'])
-                    if wd > 0.0:
-                        pp.data.mul_(1.0 - lr * wd)
-
-                if sharded and self._rs_futures[i] is not None:
-                    self._rs_futures[i].wait()
-                    g = m['shard']
-                    buf = m['shard_mom']
-                else:
-                    g = p.grad.bfloat16()
-                    state = self.state[p]
-                    if "momentum_buffer" not in state:
-                        state["momentum_buffer"] = torch.zeros_like(g)
-                    buf = state["momentum_buffer"]
-
-                buf.mul_(momentum).add_(g)
-                if nesterov:
-                    update = g.add(buf, alpha=momentum)
-                else:
-                    update = buf
-
-                update = zeropower_via_newtonschulz5(update, steps=backend_steps)
-                update = _post_ns_normalize(update, post_norm)
-
-                if sharded:
-                    prev_ag_handle = dist.all_gather_into_tensor(
-                        m['full_update'], update, async_op=True)
-                    prev_m = m
-                else:
-                    p.add_(update.to(dtype=p.dtype), alpha=-lr * m['scale'])
-                    if wd > 0.0:
-                        p.data.mul_(1.0 - lr * wd)
-
-            if prev_ag_handle is not None:
-                prev_ag_handle.wait()
-                pp = prev_m['p']
-                upd = prev_m['full_update'][:prev_m['B']]
-                pp.add_(upd.to(dtype=pp.dtype), alpha=-lr * prev_m['scale'])
-                if wd > 0.0:
-                    pp.data.mul_(1.0 - lr * wd)
-
-            if hasattr(self, '_rs_futures'):
-                del self._rs_futures
-
-        return loss
-
-# --- Tokenizer evaluation helpers ---
-
-def build_sentencepiece_luts(
-    sp: spm.SentencePieceProcessor, vocab_size: int, device: torch.device
-) -> tuple[Tensor, Tensor, Tensor]:
-    sp_vocab_size = int(sp.vocab_size())
-    table_size = max(sp_vocab_size, vocab_size)
-    base_bytes_np = np.zeros((table_size,), dtype=np.int16)
-    has_leading_space_np = np.zeros((table_size,), dtype=np.bool_)
-    is_boundary_token_np = np.ones((table_size,), dtype=np.bool_)
-    for token_id in range(sp_vocab_size):
-        if sp.is_control(token_id) or sp.is_unknown(token_id) or sp.is_unused(token_id):
-            continue
-        is_boundary_token_np[token_id] = False
-        if sp.is_byte(token_id):
-            base_bytes_np[token_id] = 1
-            continue
-        piece = sp.id_to_piece(token_id)
-        if piece.startswith("\u2581"):
-            has_leading_space_np[token_id] = True
-            piece = piece[1:]
-        base_bytes_np[token_id] = len(piece.encode("utf-8"))
-    return (
-        torch.tensor(base_bytes_np, dtype=torch.int16, device=device),
-        torch.tensor(has_leading_space_np, dtype=torch.bool, device=device),
-        torch.tensor(is_boundary_token_np, dtype=torch.bool, device=device),
-    )
-def load_validation_tokens(pattern: str, seq_len: int) -> Tensor:
-    files = [Path(p) for p in sorted(glob.glob(pattern))]
-    if not files:
-        raise FileNotFoundError(f"No files found for pattern: {pattern}")
-    tokens = torch.cat([load_data_shard(file) for file in files]).contiguous()
-    usable = ((tokens.numel() - 1) // seq_len) * seq_len
-    if usable <= 0:
-        raise ValueError(f"Validation split is too short for TRAIN_SEQ_LEN={seq_len}")
-    return tokens[: usable + 1]
-
-def eval_val(
-    args: Hyperparameters,
-    model: nn.Module,
-    rank: int,
-    world_size: int,
-    device: torch.device,
-    grad_accum_steps: int,
-    val_tokens: Tensor,
-    base_bytes_lut: Tensor,
-    has_leading_space_lut: Tensor,
-    is_boundary_token_lut: Tensor,
-    eval_seq_len: int | None = None,
-) -> tuple[float, float]:
-    seq_len = eval_seq_len or args.train_seq_len
-    local_batch_tokens = args.val_batch_size // (world_size * grad_accum_steps)
-    if local_batch_tokens < seq_len:
-        raise ValueError(
-            "VAL_BATCH_SIZE must provide at least one sequence per rank; "
-            f"got VAL_BATCH_SIZE={args.val_batch_size}, WORLD_SIZE={world_size}, "
-            f"GRAD_ACCUM_STEPS={grad_accum_steps}, seq_len={seq_len}"
-        )
-    local_batch_seqs = local_batch_tokens // seq_len
-    total_seqs = (val_tokens.numel() - 1) // seq_len
-    seq_start = (total_seqs * rank) // world_size
-    seq_end = (total_seqs * (rank + 1)) // world_size
-    val_loss_sum = torch.zeros((), device=device, dtype=torch.float64)
-    val_token_count = torch.zeros((), device=device, dtype=torch.float64)
-    val_byte_count = torch.zeros((), device=device, dtype=torch.float64)
-    model.eval()
-    with torch.inference_mode():
-        for batch_seq_start in range(seq_start, seq_end, local_batch_seqs):
-            batch_seq_end = min(batch_seq_start + local_batch_seqs, seq_end)
-            raw_start = batch_seq_start * seq_len
-            raw_end = batch_seq_end * seq_len + 1
-            local = val_tokens[raw_start:raw_end].to(device=device, dtype=torch.int64, non_blocking=True)
-            x = local[:-1].reshape(-1, seq_len)
-            y = local[1:].reshape(-1, seq_len)
-            with torch.autocast(device_type="cuda", dtype=torch.bfloat16, enabled=True):
-                batch_loss = model(x, y).detach()
-            batch_token_count = float(y.numel())
-            val_loss_sum += batch_loss.to(torch.float64) * batch_token_count
-            val_token_count += batch_token_count
-            prev_ids = x.reshape(-1)
-            tgt_ids = y.reshape(-1)
-            token_bytes = base_bytes_lut[tgt_ids].to(dtype=torch.int16)
-            token_bytes += (has_leading_space_lut[tgt_ids] & ~is_boundary_token_lut[prev_ids]).to(dtype=torch.int16)
-            val_byte_count += token_bytes.to(torch.float64).sum()
-    if dist.is_available() and dist.is_initialized():
-        dist.all_reduce(val_loss_sum, op=dist.ReduceOp.SUM)
-        dist.all_reduce(val_token_count, op=dist.ReduceOp.SUM)
-        dist.all_reduce(val_byte_count, op=dist.ReduceOp.SUM)
-    val_loss = val_loss_sum / val_token_count
-    bits_per_token = val_loss.item() / math.log(2.0)
-    tokens_per_byte = val_token_count.item() / val_byte_count.item()
-    model.train()
-    return float(val_loss.item()), float(bits_per_token * tokens_per_byte)
-
-# --- Quantization helpers ---
-
-CONTROL_TENSOR_NAME_PATTERNS = tuple(
-    pattern
-    for pattern in os.environ.get(
-        "CONTROL_TENSOR_NAME_PATTERNS",
-        "attn_scale,attn_scales,mlp_scale,mlp_scales,resid_mix,resid_mixes,q_gain,skip_weight,skip_weights,skip_gate,skip_gates,smear,ve_layer_scales,ve_shared.scale,ngram_gate",
-    ).split(",")
-    if pattern
-)
-INT8_KEEP_FLOAT_FP32_NAME_PATTERNS = tuple(
-    pattern
-    for pattern in os.environ.get(
-        "INT8_KEEP_FLOAT_FP32_NAME_PATTERNS",
-        ",".join(CONTROL_TENSOR_NAME_PATTERNS),
-    ).split(",")
-    if pattern
-)
-INT8_KEEP_FLOAT_MAX_NUMEL = 65_536
-INT8_KEEP_FLOAT_STORE_DTYPE = torch.float16
-INT8_PER_ROW_SCALE_DTYPE = torch.float16
-INT8_CLIP_PERCENTILE = 99.99984
-INT8_CLIP_Q = INT8_CLIP_PERCENTILE / 100.0
-def tensor_nbytes(t: Tensor) -> int:
-    return int(t.numel()) * int(t.element_size())
-def keep_float_tensor(name: str, t: Tensor, passthrough_orig_dtypes: dict[str, str]) -> Tensor:
-    if any(pattern in name for pattern in INT8_KEEP_FLOAT_FP32_NAME_PATTERNS):
-        return t.float().contiguous()
-    if t.dtype in {torch.float32, torch.bfloat16}:
-        passthrough_orig_dtypes[name] = str(t.dtype).removeprefix("torch.")
-        return t.to(dtype=INT8_KEEP_FLOAT_STORE_DTYPE).contiguous()
-    return t
-def quantize_float_tensor(t: Tensor) -> tuple[Tensor, Tensor]:
-    t32 = t.float()
-    if t32.ndim == 2:
-        clip_abs = (
-            torch.quantile(t32.abs(), INT8_CLIP_Q, dim=1)
-            if t32.numel()
-            else torch.empty((t32.shape[0],), dtype=torch.float32)
-        )
-        clipped = torch.maximum(torch.minimum(t32, clip_abs[:, None]), -clip_abs[:, None])
-        scale = (clip_abs / 127.0).clamp_min(1.0 / 127.0)
-        q = torch.clamp(torch.round(clipped / scale[:, None]), -127, 127).to(torch.int8).contiguous()
-        return q, scale.to(dtype=INT8_PER_ROW_SCALE_DTYPE).contiguous()
-    clip_abs = float(torch.quantile(t32.abs().flatten(), INT8_CLIP_Q).item()) if t32.numel() else 0.0
-    scale = torch.tensor(clip_abs / 127.0 if clip_abs > 0 else 1.0, dtype=torch.float32)
-    q = torch.clamp(torch.round(torch.clamp(t32, -clip_abs, clip_abs) / scale), -127, 127).to(torch.int8).contiguous()
-    return q, scale
-def quantize_state_dict_int8(state_dict: dict[str, Tensor]):
-    quantized: dict[str, Tensor] = {}
-    scales: dict[str, Tensor] = {}
-    dtypes: dict[str, str] = {}
-    passthrough: dict[str, Tensor] = {}
-    passthrough_orig_dtypes: dict[str, str] = {}
-    qmeta: dict[str, dict[str, object]] = {}
-    stats = dict.fromkeys(
-        ("param_count", "num_tensors", "num_float_tensors", "num_nonfloat_tensors", "baseline_tensor_bytes", "int8_payload_bytes"),
-        0,
-    )
-    for name, tensor in state_dict.items():
-        t = tensor.detach().to("cpu").contiguous()
-        stats["param_count"] += int(t.numel())
-        stats["num_tensors"] += 1
-        stats["baseline_tensor_bytes"] += tensor_nbytes(t)
-        if not t.is_floating_point():
-            stats["num_nonfloat_tensors"] += 1
-            passthrough[name] = t
-            stats["int8_payload_bytes"] += tensor_nbytes(t)
-            continue
-        # Keep tok_emb.weight in fp16 passthrough (tied embeddings — quantization errors degrade all tokens)
-        if t.numel() <= INT8_KEEP_FLOAT_MAX_NUMEL or name == "tok_emb.weight":
-            kept = keep_float_tensor(name, t, passthrough_orig_dtypes)
-            passthrough[name] = kept
-            stats["int8_payload_bytes"] += tensor_nbytes(kept)
-            continue
-        stats["num_float_tensors"] += 1
-        q, s = quantize_float_tensor(t)
-        if s.ndim > 0:
-            qmeta[name] = {"scheme": "per_row", "axis": 0}
-        quantized[name] = q
-        scales[name] = s
-        dtypes[name] = str(t.dtype).removeprefix("torch.")
-        stats["int8_payload_bytes"] += tensor_nbytes(q) + tensor_nbytes(s)
-    obj: dict[str, object] = {
-        "__quant_format__": "int8_clean_per_row_v1",
-        "quantized": quantized,
-        "scales": scales,
-        "dtypes": dtypes,
-        "passthrough": passthrough,
-    }
-    if qmeta:
-        obj["qmeta"] = qmeta
-    if passthrough_orig_dtypes:
-        obj["passthrough_orig_dtypes"] = passthrough_orig_dtypes
-    return obj, stats
-def dequantize_state_dict_int8(obj: dict[str, object]) -> dict[str, Tensor]:
-    out: dict[str, Tensor] = {}
-    qmeta = obj.get("qmeta", {})
-    passthrough_orig_dtypes = obj.get("passthrough_orig_dtypes", {})
-    for name, q in obj["quantized"].items():
-        dtype = getattr(torch, obj["dtypes"][name])
-        s = obj["scales"][name]
-        if qmeta.get(name, {}).get("scheme") == "per_row" or s.ndim > 0:
-            s = s.to(dtype=torch.float32)
-            out[name] = (q.float() * s.view(q.shape[0], *([1] * (q.ndim - 1)))).to(dtype=dtype).contiguous()
-        else:
-            scale = float(s.item())
-            out[name] = (q.float() * scale).to(dtype=dtype).contiguous()
-    for name, t in obj["passthrough"].items():
-        out_t = t.detach().to("cpu").contiguous()
-        orig_dtype = passthrough_orig_dtypes.get(name)
-        if isinstance(orig_dtype, str):
-            out_t = out_t.to(dtype=getattr(torch, orig_dtype)).contiguous()
-        out[name] = out_t
-    return out
-
-# --- Data loading ---
-
-def load_data_shard(file: Path) -> Tensor:
-    header_bytes = 256 * np.dtype("<i4").itemsize
-    token_bytes = np.dtype("<u2").itemsize
-    header = np.fromfile(file, dtype="<i4", count=256)
-    if header.size != 256 or int(header[0]) != 20240520 or int(header[1]) != 1:
-        raise ValueError(f"Unexpected shard header for {file}")
-    num_tokens = int(header[2])
-    expected_size = header_bytes + num_tokens * token_bytes
-    if file.stat().st_size != expected_size:
-        raise ValueError(f"Shard size mismatch for {file}: expected {expected_size} bytes")
-    tokens_np = np.fromfile(file, dtype="<u2", count=num_tokens, offset=header_bytes)
-    if tokens_np.size != num_tokens:
-        raise ValueError(f"Short read for {file}")
-    return torch.from_numpy(tokens_np.astype(np.uint16, copy=False))
-class TokenStream:
-    def __init__(self, pattern: str):
-        self.files = [Path(p) for p in sorted(glob.glob(pattern))]
-        if not self.files:
-            raise FileNotFoundError(f"No files found for pattern: {pattern}")
-        self.file_idx = 0
-        self.tokens = load_data_shard(self.files[0])
-        self.pos = 0
-    def _advance_file(self) -> None:
-        self.file_idx = (self.file_idx + 1) % len(self.files)
-        self.tokens = load_data_shard(self.files[self.file_idx])
-        self.pos = 0
-    def take(self, n: int) -> Tensor:
-        chunks: list[Tensor] = []
-        remaining = n
-        while remaining > 0:
-            avail = self.tokens.numel() - self.pos
-            if avail <= 0:
-                self._advance_file()
-                continue
-            k = min(remaining, avail)
-            chunks.append(self.tokens[self.pos : self.pos + k])
-            self.pos += k
-            remaining -= k
-        return chunks[0] if len(chunks) == 1 else torch.cat(chunks)
-class DistributedTokenLoader:
-    def __init__(self, pattern: str, rank: int, world_size: int, device: torch.device):
-        self.rank = rank
-        self.world_size = world_size
-        self.device = device
-        self.stream = TokenStream(pattern)
-    def next_batch(self, global_tokens: int, seq_len: int, grad_accum_steps: int) -> tuple[Tensor, Tensor]:
-        local_tokens = global_tokens // (self.world_size * grad_accum_steps)
-        per_rank_span = local_tokens + 1
-        chunk = self.stream.take(per_rank_span * self.world_size)
-        start = self.rank * per_rank_span
-        local = chunk[start : start + per_rank_span].to(dtype=torch.int64)
-        x = local[:-1].reshape(-1, seq_len)
-        y = local[1:].reshape(-1, seq_len)
-        return x.to(self.device, non_blocking=True), y.to(self.device, non_blocking=True)
-
-# --- Transformer modules ---
-
-class RMSNorm(nn.Module):
-    def __init__(self, eps: float | None = None):
-        super().__init__()
-        self.eps = eps
-    def forward(self, x: Tensor) -> Tensor:
-        return F.rms_norm(x, (x.size(-1),), eps=self.eps)
-
-class CastedLinear(nn.Linear):
-    _qat_enabled: bool = False
-    _qat_clip_pct: float = 0.9995
-    _qat_default_bits: int = 5
-    _qat_soft_round: bool = False
-    _qat_soft_alpha: Tensor = None  # type: ignore[assignment]  # CUDA tensor, ramped 1→16; None until QAT starts
-
-    def forward(self, x: Tensor) -> Tensor:
-        w = self.weight.to(x.dtype)
-        if CastedLinear._qat_enabled and self.training and w.ndim == 2:
-            bits = getattr(self, '_qat_bits', CastedLinear._qat_default_bits)
-            if CastedLinear._qat_soft_round:
-                w = _apply_qat_soft_round(w, self.weight, bits, CastedLinear._qat_soft_alpha)
-            else:
-                w = _apply_qat_ste(w, self.weight, bits)
-        bias = self.bias.to(x.dtype) if self.bias is not None else None
-        return F.linear(x, w, bias)
-
-
-def _apply_qat_ste(w_cast: Tensor, w_fp32: Tensor, bits: int) -> Tensor:
-    """Straight-through estimator QAT for a weight matrix."""
-    if bits <= 0:
-        return w_cast
-    qmax = (1 << (bits - 1)) - 1
-    qmin = -(1 << (bits - 1))
-    with torch.no_grad():
-        w32 = w_fp32.float()
-        row_max = w32.abs().amax(dim=-1) * CastedLinear._qat_clip_pct
-        scale = (row_max / float(qmax)).clamp_min(1.0 / float(qmax))
-        w_q = (torch.clamp(torch.round(w32 / scale.unsqueeze(-1)), qmin, qmax) * scale.unsqueeze(-1)).to(w_cast.dtype)
-    return w_cast + (w_q - w_cast).detach()
-
-
-def _apply_qat_soft_round(w_cast: Tensor, w_fp32: Tensor, bits: int, alpha: "float | Tensor") -> Tensor:
-    """Differentiable quantization using soft-round (sigmoid approximation of round()).
-
-    Instead of STE (zero gradient through round), uses sigmoid(alpha*(frac-0.5)) which
-    provides real gradient signal pushing weights toward quantization grid points.
-    alpha ramps from 1 (smooth) to 16 (near-hard rounding) during QAT.
-    """
-    if bits <= 0:
-        return w_cast
-    out_dtype = w_cast.dtype
-    qmax = (1 << (bits - 1)) - 1
-    qmin = -(1 << (bits - 1))
-    # Scale: detached so gradient doesn't flow through row_max computation
-    row_max = w_fp32.float().detach().abs().amax(dim=-1) * CastedLinear._qat_clip_pct
-    scale = (row_max / float(qmax)).clamp_min(1.0 / float(qmax))
-    # Differentiable quantization
-    w_scaled = w_fp32.float() / scale.unsqueeze(-1)
-    w_clamped = w_scaled.clamp(float(qmin), float(qmax))
-    w_floor = w_clamped.detach().floor()  # integer part (no grad)
-    frac = w_clamped - w_floor  # fractional part (grad flows through w_clamped)
-    soft = w_floor + torch.sigmoid(alpha * (frac - 0.5))
-    w_q = soft * scale.unsqueeze(-1)
-    return w_q.to(out_dtype)
-
-
-def _apply_bank_qat(w: Tensor, bits: int, dtype: torch.dtype) -> Tensor:
-    """Apply QAT to a bank weight slice, returning the result in the target dtype."""
-    w_cast = w.to(dtype)
-    if CastedLinear._qat_enabled and torch.is_grad_enabled() and w.ndim == 2:
-        if CastedLinear._qat_soft_round:
-            return _apply_qat_soft_round(w_cast, w, bits, CastedLinear._qat_soft_alpha)
-        return _apply_qat_ste(w_cast, w, bits)
-    return w_cast
-
-
-def restore_low_dim_params_to_fp32(module: nn.Module) -> None:
-    with torch.no_grad():
-        for name, param in module.named_parameters():
-            if (param.ndim < 2 or any(pattern in name for pattern in CONTROL_TENSOR_NAME_PATTERNS)) and param.dtype != torch.float32:
-                param.data = param.data.float()
-class Rotary(nn.Module):
-    def __init__(self, dim: int, base: float = 10000.0, train_seq_len: int = 1024, rope_dims: int = 0):
-        super().__init__()
-        self.dim = dim
-        self.base = base
-        self.train_seq_len = train_seq_len
-        self.rope_dims = rope_dims if rope_dims > 0 else dim
-        inv_freq = 1.0 / (base ** (torch.arange(0, self.rope_dims, 2, dtype=torch.float32) / self.rope_dims))
-        self.register_buffer("inv_freq", inv_freq, persistent=False)
-        self._seq_len_cached = 0
-        self._cos_cached: Tensor | None = None
-        self._sin_cached: Tensor | None = None
-    def forward(self, seq_len: int, device: torch.device, dtype: torch.dtype) -> tuple[Tensor, Tensor]:
-        if (
-            self._cos_cached is None
-            or self._sin_cached is None
-            or self._seq_len_cached != seq_len
-            or self._cos_cached.device != device
-        ):
-            rd = self.rope_dims
-            if seq_len > self.train_seq_len:
-                scale = seq_len / self.train_seq_len
-                new_base = self.base * (scale ** (rd / (rd - 2)))
-                inv_freq = 1.0 / (new_base ** (torch.arange(0, rd, 2, dtype=torch.float32, device=device) / rd))
-            else:
-                inv_freq = self.inv_freq.to(device)
-            t = torch.arange(seq_len, device=device, dtype=inv_freq.dtype)
-            freqs = torch.outer(t, inv_freq)
-            self._cos_cached = freqs.cos()[None, :, None, :]
-            self._sin_cached = freqs.sin()[None, :, None, :]
-            self._seq_len_cached = seq_len
-        return self._cos_cached.to(dtype=dtype), self._sin_cached.to(dtype=dtype)
-def apply_rotary_emb(x: Tensor, cos: Tensor, sin: Tensor, rope_dims: int = 0) -> Tensor:
-    if rope_dims > 0 and rope_dims < x.size(-1):
-        x_rope, x_pass = x[..., :rope_dims], x[..., rope_dims:]
-        half = rope_dims // 2
-        x1, x2 = x_rope[..., :half], x_rope[..., half:]
-        x_rope = torch.cat((x1 * cos + x2 * sin, x1 * (-sin) + x2 * cos), dim=-1)
-        return torch.cat((x_rope, x_pass), dim=-1)
-    half = x.size(-1) // 2
-    x1, x2 = x[..., :half], x[..., half:]
-    return torch.cat((x1 * cos + x2 * sin, x1 * (-sin) + x2 * cos), dim=-1)
-
-class CausalSelfAttention(nn.Module):
-    def __init__(
-        self,
-        dim: int,
-        num_heads: int,
-        num_kv_heads: int,
-        rope_base: float,
-        qk_gain_init: float,
-    ):
-        super().__init__()
-        if dim % num_heads != 0:
-            raise ValueError("model_dim must be divisible by num_heads")
-        if num_heads % num_kv_heads != 0:
-            raise ValueError("num_heads must be divisible by num_kv_heads")
-        self.num_heads = num_heads
-        self.num_kv_heads = num_kv_heads
-        self.head_dim = dim // num_heads
-        if self.head_dim % 2 != 0:
-            raise ValueError("head_dim must be even for RoPE")
-        # No CastedLinear -- weights come from banks
-        self.q_gain = nn.Parameter(torch.full((num_heads,), qk_gain_init, dtype=torch.float32))
-        self.rope_dims = 0  # set by GPT.__init__ for partial RoPE
-        self.rotary = Rotary(self.head_dim, base=rope_base, train_seq_len=1024)
-        self.use_xsa = False  # set by GPT.__init__ for deep layers only
-    def _xsa_efficient(self, y: Tensor, v: Tensor) -> Tensor:
-        """Efficient XSA: subtract self-value projection via GQA-aware reshape (no repeat_interleave).
-        y: [B, T, H, D], v: [B, T, Hkv, D]. H must be divisible by Hkv."""
-        B, T, H, D = y.shape
-        Hkv = v.size(-2)
-        group = H // Hkv
-        y_g = y.reshape(B, T, Hkv, group, D)        # [B, T, Hkv, group, D]
-        vn = F.normalize(v, dim=-1).unsqueeze(-2)    # [B, T, Hkv, 1, D] -- broadcast ready
-        proj = (y_g * vn).sum(dim=-1, keepdim=True) * vn
-        return (y_g - proj).reshape(B, T, H, D)
-    def forward(self, x: Tensor, q_w: Tensor, k_w: Tensor, v_w: Tensor, out_w: Tensor, v_embed: Tensor | None = None) -> tuple[Tensor, Tensor | None]:
-        bsz, seqlen, dim = x.shape
-        _qb = CastedLinear._qat_default_bits
-        q = F.linear(x, _apply_bank_qat(q_w, _qb, x.dtype)).reshape(bsz, seqlen, self.num_heads, self.head_dim)
-        k = F.linear(x, _apply_bank_qat(k_w, _qb, x.dtype)).reshape(bsz, seqlen, self.num_kv_heads, self.head_dim)
-        v = F.linear(x, _apply_bank_qat(v_w, _qb, x.dtype))
-        if v_embed is not None:
-            v = v + v_embed
-        v = v.reshape(bsz, seqlen, self.num_kv_heads, self.head_dim)
-        q = F.rms_norm(q, (q.size(-1),))
-        k = F.rms_norm(k, (k.size(-1),))
-        cos, sin = self.rotary(seqlen, x.device, q.dtype)
-        q = apply_rotary_emb(q, cos, sin, self.rope_dims)
-        k = apply_rotary_emb(k, cos, sin, self.rope_dims)
-        q = q * self.q_gain.to(dtype=q.dtype)[None, None, :, None]
-        # SDPA expects (B, H, T, D)
-        y = F.scaled_dot_product_attention(
-            q.transpose(1, 2), k.transpose(1, 2), v.transpose(1, 2),
-            is_causal=True, enable_gqa=(self.num_kv_heads != self.num_heads),
-        ).transpose(1, 2)
-        if self.use_xsa:
-            y = self._xsa_efficient(y, v)
-        y = y.reshape(bsz, seqlen, dim)
-        return F.linear(y, _apply_bank_qat(out_w, _qb, y.dtype)), None
-
-class SmearGate(nn.Module):
-    def __init__(self, dim: int):
-        super().__init__()
-        self.gate = nn.Parameter(torch.zeros(dim, dtype=torch.float32))
-    def forward(self, x: Tensor) -> Tensor:
-        g = torch.sigmoid(self.gate.to(dtype=x.dtype))[None, None, :]
-        x_prev = F.pad(x[:, :-1], (0, 0, 1, 0))  # prepend zero row, avoids cat+alloc
-        return torch.lerp(x, x_prev, g)
-
-class EngramLite(nn.Module):
-    """Multi-head hash-based n-gram embedding with learned gating (Engram-lite).
-
-    Replaces BigramHashEmbedding with: multi-head hashing for collision resistance,
-    bigram+trigram coverage, and a per-dim learned gate to suppress n-gram signal
-    when the Transformer's own reasoning is more informative.
-    """
-    def __init__(self, num_buckets: int, num_heads: int, num_orders: int, dim_per_head: int, model_dim: int):
-        super().__init__()
-        self.num_buckets = num_buckets
-        self.num_heads = num_heads
-        self.num_orders = num_orders
-        self.dim_per_head = dim_per_head
-        total_slots = num_orders * num_heads * num_buckets
-        concat_dim = num_orders * num_heads * dim_per_head
-        self.embed = nn.Embedding(total_slots, dim_per_head)
-        nn.init.normal_(self.embed.weight, std=0.01)
-        self.proj = CastedLinear(concat_dim, model_dim, bias=False)
-        self.proj._zero_init = True
-        self.ngram_gate = nn.Parameter(torch.zeros(model_dim, dtype=torch.float32))
-
-    def forward(self, input_ids: Tensor) -> Tensor:
-        B = self.num_buckets
-        prev_ids = F.pad(input_ids[:, :-1], (1, 0), value=0)
-        # Bigram hashes (2 heads, independent prime-based mixing)
-        bi_h0 = (prev_ids * 1009 + input_ids) % B
-        bi_h1 = ((prev_ids * 2719 + 314159) ^ (input_ids * 3137)) % B
-        indices = [bi_h0, bi_h1 + B]
-        # Trigram hashes (2 heads) if enabled
-        if self.num_orders >= 2:
-            pp_ids = F.pad(prev_ids[:, :-1], (1, 0), value=0)
-            tri_h0 = ((pp_ids * 36313) ^ (prev_ids * 27191) ^ (input_ids * 4903)) % B
-            tri_h1 = ((pp_ids * 7919) ^ (prev_ids * 4391) ^ (input_ids * 6151)) % B
-            offset = 2 * B
-            indices.extend([tri_h0 + offset, tri_h1 + offset + B])
-        # Unified lookup + concat
-        all_idx = torch.stack(indices, dim=-1)
-        all_emb = self.embed(all_idx)
-        flat = all_emb.reshape(*input_ids.shape, -1)
-        out = self.proj(flat)
-        gate = torch.sigmoid(self.ngram_gate.to(dtype=out.dtype))[None, None, :]
-        return out * gate
-
-class ValueEmbedding(nn.Module):
-    """Reinject token identity into attention values at specific layers.
-    Each table maps vocab tokens to a low-dim embedding, projected to kv_dim."""
-    def __init__(self, vocab_size: int, ve_dim: int, kv_dim: int):
-        super().__init__()
-        self.embed = nn.Embedding(vocab_size, ve_dim)
-        nn.init.normal_(self.embed.weight, std=0.01)
-        self.proj = CastedLinear(ve_dim, kv_dim, bias=False) if ve_dim != kv_dim else None
-        if self.proj is not None:
-            nn.init.zeros_(self.proj.weight)
-        self.scale = nn.Parameter(torch.tensor(0.1, dtype=torch.float32))
-    def forward(self, token_ids: Tensor) -> Tensor:
-        h = self.embed(token_ids)
-        if self.proj is not None:
-            h = self.proj(h)
-        return h * self.scale.to(dtype=h.dtype)
-
-class MLP(nn.Module):
-    def __init__(self, dim: int, mlp_mult: float):
-        super().__init__()
-        # No CastedLinear -- weights come from banks
-    def forward(self, x: Tensor, up_w: Tensor, down_w: Tensor) -> Tensor:
-        _qb = CastedLinear._qat_default_bits
-        x = F.leaky_relu(F.linear(x, _apply_bank_qat(up_w, _qb, x.dtype)), negative_slope=0.3)
-        return F.linear(x.square(), _apply_bank_qat(down_w, _qb, x.dtype))
-
-class Block(nn.Module):
-    def __init__(
-        self,
-        dim: int,
-        num_heads: int,
-        num_kv_heads: int,
-        mlp_mult: float,
-        rope_base: float,
-        qk_gain_init: float,
-        layer_idx: int = 0,
-        ln_scale: bool = False,
-    ):
-        super().__init__()
-        self.attn_norm = RMSNorm()
-        self.mlp_norm = RMSNorm()
-        self.attn = CausalSelfAttention(dim, num_heads, num_kv_heads, rope_base, qk_gain_init)
-        self.mlp = MLP(dim, mlp_mult)
-        self.attn_scale = nn.Parameter(torch.ones(dim, dtype=torch.float32))
-        self.mlp_scale = nn.Parameter(torch.ones(dim, dtype=torch.float32))
-        self.resid_mix = nn.Parameter(torch.stack((torch.ones(dim), torch.zeros(dim))).float())
-        self.ln_scale_factor = 1.0 / math.sqrt(layer_idx + 1) if ln_scale else 1.0
-    def forward(self, x: Tensor, x0: Tensor, q_w: Tensor, k_w: Tensor, v_w: Tensor, out_w: Tensor, up_w: Tensor, down_w: Tensor, v_embed: Tensor | None = None) -> tuple[Tensor, Tensor | None]:
-        mix = self.resid_mix.to(dtype=x.dtype)
-        x_in = mix[0][None, None, :] * x + mix[1][None, None, :] * x0
-        attn_out, raw_v = self.attn(self.attn_norm(x_in) * self.ln_scale_factor, q_w, k_w, v_w, out_w, v_embed=v_embed)
-        x_out = x_in + self.attn_scale.to(dtype=x_in.dtype)[None, None, :] * attn_out
-        x_out = x_out + self.mlp_scale.to(dtype=x_out.dtype)[None, None, :] * self.mlp(self.mlp_norm(x_out) * self.ln_scale_factor, up_w, down_w)
-        return x_out, raw_v
-
-class GPT(nn.Module):
-    def __init__(
-        self,
-        vocab_size: int,
-        num_layers: int,
-        model_dim: int,
-        num_heads: int,
-        num_kv_heads: int,
-        mlp_mult: float,
-        tie_embeddings: bool,
-        tied_embed_init_std: float,
-        logit_softcap: float,
-        rope_base: float,
-        qk_gain_init: float,
-        ngram_buckets: int = 0,
-        ngram_heads: int = 2,
-        ngram_orders: int = 2,
-        ngram_dim_per_head: int = 32,
-        xsa_last_n: int = 0,
-        rope_dims: int = 0,
-        ln_scale: bool = False,
-        ve_enabled: bool = False,
-        ve_dim: int = 128,
-        ve_layers: str = "9,10",
-    ):
-        super().__init__()
-        self._ve_target_dim = num_kv_heads * (model_dim // num_heads)  # kv_dim for value projection
-        if logit_softcap <= 0.0:
-            raise ValueError(f"logit_softcap must be positive, got {logit_softcap}")
-        self.tie_embeddings = tie_embeddings
-        self.tied_embed_init_std = tied_embed_init_std
-        self.logit_softcap = logit_softcap
-        self.tok_emb = nn.Embedding(vocab_size, model_dim)
-        self.bigram = EngramLite(ngram_buckets, ngram_heads, ngram_orders, ngram_dim_per_head, model_dim) if ngram_buckets > 0 else None
-        self.smear = SmearGate(model_dim)
-        self.num_encoder_layers = num_layers // 2
-        self.num_decoder_layers = num_layers - self.num_encoder_layers
-        self.num_skip_weights = min(self.num_encoder_layers, self.num_decoder_layers)
-        self.skip_weights = nn.Parameter(torch.ones(self.num_skip_weights, model_dim, dtype=torch.float32))
-        self.skip_gates = nn.Parameter(torch.zeros(self.num_skip_weights, model_dim, dtype=torch.float32))
-        # Parameter banks: contiguous 3D tensors for batched optimizer
-        head_dim = model_dim // num_heads
-        kv_dim = num_kv_heads * head_dim
-        mlp_dim = int(mlp_mult * model_dim)
-        self.num_layers = num_layers
-        self.qo_bank = nn.Parameter(torch.empty(2 * num_layers, model_dim, model_dim))
-        self.kv_bank = nn.Parameter(torch.empty(2 * num_layers, kv_dim, model_dim))
-        self.mlp_up_bank = nn.Parameter(torch.empty(num_layers, mlp_dim, model_dim))
-        self.mlp_down_bank = nn.Parameter(torch.empty(num_layers, model_dim, mlp_dim))
-        self.blocks = nn.ModuleList(
-            [
-                Block(
-                    model_dim,
-                    num_heads,
-                    num_kv_heads,
-                    mlp_mult,
-                    rope_base,
-                    qk_gain_init,
-                    layer_idx=i,
-                    ln_scale=ln_scale,
-                )
-                for i in range(num_layers)
-            ]
-        )
-        if rope_dims > 0:
-            head_dim = model_dim // num_heads
-            for block in self.blocks:
-                block.attn.rope_dims = rope_dims
-                block.attn.rotary = Rotary(head_dim, base=rope_base, train_seq_len=1024, rope_dims=rope_dims)
-        self.ve_layer_indices = [int(x) for x in ve_layers.split(",") if x.strip()] if ve_enabled else []
-        kv_dim_ve = self._ve_target_dim
-        if self.ve_layer_indices:
-            self.ve_shared = ValueEmbedding(vocab_size, ve_dim, kv_dim_ve)
-            self.ve_layer_scales = nn.ParameterList(
-                [nn.Parameter(torch.ones(1, dtype=torch.float32)) for _ in self.ve_layer_indices]
-            )
-        else:
-            self.ve_shared = None
-            self.ve_layer_scales = nn.ParameterList()
-        self.value_embeds = nn.ModuleList()  # keep empty for compat
-        self.final_norm = RMSNorm()
-        self.lm_head = None if tie_embeddings else CastedLinear(model_dim, vocab_size, bias=False)
-        if self.lm_head is not None:
-            self.lm_head._zero_init = True
-        if xsa_last_n > 0:
-            for i in range(max(0, num_layers - xsa_last_n), num_layers):
-                self.blocks[i].attn.use_xsa = True
-        self._init_weights()
-    def _init_weights(self) -> None:
-        if self.tie_embeddings:
-            nn.init.normal_(self.tok_emb.weight, mean=0.0, std=self.tied_embed_init_std)
-        n = self.num_layers
-        mimetic_alpha = 0.05
-        head_dim = self.qo_bank.shape[1] // (self.blocks[0].attn.num_heads)  # model_dim // num_heads
-        num_kv_heads = self.blocks[0].attn.num_kv_heads
-        num_heads = self.blocks[0].attn.num_heads
-        group = num_heads // num_kv_heads
-        # Init banks: orthogonal, with proj layers scaled down and out zero-init via mimetic V-O
-        for i in range(n):
-            nn.init.orthogonal_(self.qo_bank.data[i], gain=1.0)        # Q
-            nn.init.orthogonal_(self.kv_bank.data[i], gain=1.0)        # K
-            nn.init.orthogonal_(self.kv_bank.data[n + i], gain=1.0)    # V
-            nn.init.orthogonal_(self.mlp_up_bank.data[i], gain=1.0)    # MLP up
-            nn.init.zeros_(self.mlp_down_bank.data[i])                  # MLP down (zero init)
-            # Mimetic V-O Init for output projection: O_h = -alpha * V_h per head
-            v_w = self.kv_bank.data[n + i]  # (kv_dim, model_dim)
-            o_w = torch.zeros_like(self.qo_bank.data[n + i])  # (model_dim, model_dim)
-            for kv_h in range(num_kv_heads):
-                v_block = v_w[kv_h * head_dim : (kv_h + 1) * head_dim, :]
-                for g_idx in range(group):
-                    q_h = kv_h * group + g_idx
-                    o_w[q_h * head_dim : (q_h + 1) * head_dim, :] = -mimetic_alpha * v_block
-            self.qo_bank.data[n + i].copy_(o_w)
-        # Init remaining nn.Linear modules (bigram proj, lm_head)
-        for name, module in self.named_modules():
-            if isinstance(module, nn.Linear):
-                if getattr(module, "_zero_init", False):
-                    nn.init.zeros_(module.weight)
-                elif module.weight.ndim == 2 and module.weight.shape[0] >= 64 and module.weight.shape[1] >= 64:
-                    nn.init.orthogonal_(module.weight, gain=1.0)
-    def _compute_ve_base(self, input_ids: Tensor) -> Tensor | None:
-        """Precompute shared value embedding base once per forward pass."""
-        if self.ve_shared is None:
-            return None
-        return self.ve_shared(input_ids)
-
-    def _get_ve(self, layer_idx: int, ve_base: Tensor | None) -> Tensor | None:
-        """Get value embedding for a specific layer using precomputed base + per-layer scale."""
-        if ve_base is None or layer_idx not in self.ve_layer_indices:
-            return None
-        ve_idx = self.ve_layer_indices.index(layer_idx)
-        return ve_base * self.ve_layer_scales[ve_idx].to(dtype=ve_base.dtype)
-    def forward(self, input_ids: Tensor, target_ids: Tensor) -> Tensor:
-        n = self.num_layers
-        x = self.tok_emb(input_ids)
-        if self.bigram is not None:
-            x = x + self.bigram(input_ids)
-        x = F.rms_norm(x, (x.size(-1),))
-        x = self.smear(x)
-        x0 = x
-        skips: list[Tensor] = []
-        ve_base = self._compute_ve_base(input_ids)
-        for i in range(self.num_encoder_layers):
-            ve = self._get_ve(i, ve_base)
-            x, _raw_v = self.blocks[i](x, x0,
-                self.qo_bank[i], self.kv_bank[i], self.kv_bank[n + i],
-                self.qo_bank[n + i], self.mlp_up_bank[i], self.mlp_down_bank[i],
-                v_embed=ve)
-            skips.append(x)
-        for i in range(self.num_decoder_layers):
-            bi = self.num_encoder_layers + i
-            if skips:
-                skip = skips.pop()
-                g = torch.sigmoid(self.skip_gates[i].to(dtype=x.dtype))[None, None, :]
-                scaled_skip = self.skip_weights[i].to(dtype=x.dtype)[None, None, :] * skip
-                x = torch.lerp(scaled_skip, x, g)
-            ve = self._get_ve(bi, ve_base)
-            x, _raw_v = self.blocks[bi](x, x0,
-                self.qo_bank[bi], self.kv_bank[bi], self.kv_bank[n + bi],
-                self.qo_bank[n + bi], self.mlp_up_bank[bi], self.mlp_down_bank[bi],
-                v_embed=ve)
-        x = self.final_norm(x)
-        x_flat = x.reshape(-1, x.size(-1))
-        targets = target_ids.reshape(-1)
-        if self.tie_embeddings:
-            logits_proj = F.linear(x_flat, self.tok_emb.weight)
-        else:
-            if self.lm_head is None:
-                raise RuntimeError("lm_head is required when tie_embeddings=False")
-            logits_proj = self.lm_head(x_flat)
-        logits = self.logit_softcap * torch.tanh(logits_proj / self.logit_softcap)
-        return F.cross_entropy(logits.float(), targets, reduction="mean")
-    def forward_logits(self, input_ids: Tensor) -> Tensor:
-        """Return logits (bsz, seq_len, vocab) without computing loss."""
-        n = self.num_layers
-        x = self.tok_emb(input_ids)
-        if self.bigram is not None:
-            x = x + self.bigram(input_ids)
-        x = F.rms_norm(x, (x.size(-1),))
-        x = self.smear(x)
-        x0 = x
-        skips: list[Tensor] = []
-        ve_base = self._compute_ve_base(input_ids)
-        for i in range(self.num_encoder_layers):
-            ve = self._get_ve(i, ve_base)
-            x, _raw_v = self.blocks[i](x, x0,
-                self.qo_bank[i], self.kv_bank[i], self.kv_bank[n + i],
-                self.qo_bank[n + i], self.mlp_up_bank[i], self.mlp_down_bank[i],
-                v_embed=ve)
-            skips.append(x)
-        for i in range(self.num_decoder_layers):
-            bi = self.num_encoder_layers + i
-            if skips:
-                skip = skips.pop()
-                g = torch.sigmoid(self.skip_gates[i].to(dtype=x.dtype))[None, None, :]
-                scaled_skip = self.skip_weights[i].to(dtype=x.dtype)[None, None, :] * skip
-                x = torch.lerp(scaled_skip, x, g)
-            ve = self._get_ve(bi, ve_base)
-            x, _raw_v = self.blocks[bi](x, x0,
-                self.qo_bank[bi], self.kv_bank[bi], self.kv_bank[n + bi],
-                self.qo_bank[n + bi], self.mlp_up_bank[bi], self.mlp_down_bank[bi],
-                v_embed=ve)
-        x = self.final_norm(x)
-        if self.tie_embeddings:
-            logits_proj = F.linear(x, self.tok_emb.weight)
-        else:
-            logits_proj = self.lm_head(x)
-        return self.logit_softcap * torch.tanh(logits_proj / self.logit_softcap)
-
-# --- Sliding window evaluation ---
-
-def eval_val_sliding(
-    args: Hyperparameters,
-    base_model: nn.Module,
-    rank: int,
-    world_size: int,
-    device: torch.device,
-    val_tokens: Tensor,
-    base_bytes_lut: Tensor,
-    has_leading_space_lut: Tensor,
-    is_boundary_token_lut: Tensor,
-    stride: int,
-    batch_seqs: int = 32,
-    eval_seq_len: int | None = None,
-) -> tuple[float, float]:
-    """Sliding window evaluation: each token scored with maximum context."""
-    seq_len = eval_seq_len or args.train_seq_len
-    total_tokens = val_tokens.numel() - 1
-    window_starts = [ws for ws in range(0, total_tokens, stride)
-                     if min(ws + seq_len, total_tokens) - ws >= 1]
-    total_windows = len(window_starts)
-    my_s = (total_windows * rank) // world_size
-    my_e = (total_windows * (rank + 1)) // world_size
-    my_windows = window_starts[my_s:my_e]
-    loss_sum = torch.zeros((), device=device, dtype=torch.float64)
-    token_count = torch.zeros((), device=device, dtype=torch.float64)
-    byte_count = torch.zeros((), device=device, dtype=torch.float64)
-    base_model.eval()
-    compiled_logits = torch.compile(base_model.forward_logits, dynamic=True, fullgraph=True)
-    with torch.inference_mode():
-        for bi in range(0, len(my_windows), batch_seqs):
-            batch_ws = my_windows[bi:bi + batch_seqs]
-            bsz = len(batch_ws)
-            x_batch = torch.zeros(bsz, seq_len, dtype=torch.int64, device=device)
-            y_batch = torch.zeros(bsz, seq_len, dtype=torch.int64, device=device)
-            wlens: list[int] = []
-            for i, ws in enumerate(batch_ws):
-                end = min(ws + seq_len, total_tokens)
-                wlen = end - ws
-                wlens.append(wlen)
-                chunk = val_tokens[ws:end + 1].to(dtype=torch.int64, device=device)
-                x_batch[i, :wlen] = chunk[:-1]
-                y_batch[i, :wlen] = chunk[1:]
-            with torch.autocast(device_type="cuda", dtype=torch.bfloat16):
-                logits = compiled_logits(x_batch)
-            nll = F.cross_entropy(
-                logits.reshape(-1, logits.size(-1)).float(),
-                y_batch.reshape(-1),
-                reduction="none",
-            ).reshape(bsz, seq_len)
-            for i, ws in enumerate(batch_ws):
-                wlen = wlens[i]
-                s = 0 if ws == 0 else max(wlen - stride, 0)
-                scored_nll = nll[i, s:wlen].to(torch.float64)
-                loss_sum += scored_nll.sum()
-                token_count += float(wlen - s)
-                tgt = y_batch[i, s:wlen]
-                prev = x_batch[i, s:wlen]
-                tb = base_bytes_lut[tgt].to(torch.float64)
-                tb += (has_leading_space_lut[tgt] & ~is_boundary_token_lut[prev]).to(torch.float64)
-                byte_count += tb.sum()
-    if dist.is_available() and dist.is_initialized():
-        dist.all_reduce(loss_sum, op=dist.ReduceOp.SUM)
-        dist.all_reduce(token_count, op=dist.ReduceOp.SUM)
-        dist.all_reduce(byte_count, op=dist.ReduceOp.SUM)
-    val_loss = (loss_sum / token_count).item()
-    bits_per_token = val_loss / math.log(2.0)
-    tokens_per_byte = token_count.item() / byte_count.item()
-    base_model.train()
-    return val_loss, bits_per_token * tokens_per_byte
-
-
-# --- GPTQ-lite int6 quantization ---
-
-def _classify_param(name: str) -> str:
-    if "tok_emb" in name or "lm_head" in name:
-        return "embed"
-    if ".mlp." in name:
-        return "mlp"
-    if ".attn." in name or (".proj." in name and ".mlp." not in name):
-        return "attn"
-    return "other"
-def quantize_int6_per_row(t: Tensor, clip_range: int = 31) -> tuple[Tensor, Tensor]:
-    t32 = t.float()
-    if t32.ndim == 2:
-        best_q, best_s, best_err = None, None, float('inf')
-        for pct in [0.9990, 0.9995, 0.9999, 0.99999, 1.0]:
-            if pct < 1.0:
-                row_clip = torch.quantile(t32.abs(), pct, dim=1)
-            else:
-                row_clip = t32.abs().amax(dim=1)
-            s = (row_clip / clip_range).clamp_min(1.0 / clip_range).to(torch.float16)
-            q = torch.clamp(torch.round(t32 / s.float()[:, None]), -clip_range, clip_range).to(torch.int8)
-            recon = q.float() * s.float()[:, None]
-            err = (t32 - recon).pow(2).mean().item()
-            if err < best_err:
-                best_q, best_s, best_err = q, s, err
-        return best_q, best_s
-    amax = t32.abs().max().item()
-    scale = torch.tensor(amax / clip_range if amax > 0 else 1.0, dtype=torch.float16)
-    q = torch.clamp(torch.round(t32 / scale.float()), -clip_range, clip_range).to(torch.int8)
-    return q, scale
-
-def _precompute_row_scales(W: Tensor, qmax: int) -> Tensor:
-    """Pre-compute optimal per-row scales by searching percentile clipping thresholds.
-    PR #753-style: find best scale once, then run GPTQ once with fixed scales."""
-    t32 = W.float()
-    best_s = t32.abs().amax(dim=1) / float(qmax)
-    best_s = best_s.clamp_min(1.0 / float(qmax))
-    best_err = torch.full((t32.shape[0],), float("inf"), device=t32.device)
-    for pct in [0.9990, 0.9995, 0.9999, 0.99999, 1.0]:
-        if pct < 1.0:
-            row_clip = torch.quantile(t32.abs(), pct, dim=1)
-        else:
-            row_clip = t32.abs().amax(dim=1)
-        s = (row_clip / float(qmax)).clamp_min(1.0 / float(qmax))
-        q = torch.clamp(torch.round(t32 / s[:, None]), -qmax, qmax)
-        recon = q * s[:, None]
-        err = (t32 - recon).pow(2).mean(dim=1)
-        improved = err < best_err
-        best_s[improved] = s[improved]
-        best_err[improved] = err[improved]
-    return best_s
-
-def _gptq_block_sweep(W: Tensor, Hinv: Tensor, sf: Tensor, qmin: int, qmax: int, block_size: int) -> Tensor:
-    """Run one GPTQ block-column sweep with fixed per-row scales. Returns quantized Q."""
-    rows, cols = W.shape
-    Q = torch.zeros_like(W, dtype=torch.int8)
-    W_work = W.clone()
-    for i1 in range(0, cols, block_size):
-        i2 = min(i1 + block_size, cols)
-        count = i2 - i1
-        W1 = W_work[:, i1:i2].clone()
-        Q1 = torch.zeros(rows, count, dtype=torch.int8, device=W.device)
-        Err1 = torch.zeros(rows, count, device=W.device)
-        Hinv1 = Hinv[i1:i2, i1:i2]
-        for i in range(count):
-            w = W1[:, i]
-            d = Hinv1[i, i]
-            q = torch.clamp(torch.round(w / sf), qmin, qmax).to(torch.int8)
-            Q1[:, i] = q
-            err = (w - q.float() * sf) / d
-            W1[:, i:] -= err.unsqueeze(1) * Hinv1[i, i:].unsqueeze(0)
-            Err1[:, i] = err
-        Q[:, i1:i2] = Q1
-        if i2 < cols:
-            W_work[:, i2:] -= Err1 @ Hinv[i1:i2, i2:]
-    return Q
-
-def quantize_int6_gptq(weight, hessian=None, clip_range=31, block_size=128, damp_factor=0.01,
-                        col_order="desc", single_pass=False):
-    """Full GPTQ: Hessian-aware int6 quantization with Cholesky error compensation.
-    If hessian is None, falls back to percentile search."""
-    device = hessian.device if hessian is not None else weight.device
-    t32 = weight.float().to(device)
-    if t32.ndim != 2 or hessian is None:
-        return _quantize_int6_percentile(t32, clip_range)
-    rows, cols = t32.shape
-    H = hessian.float().clone()
-    diag = torch.diag(H)
-    dead = diag == 0
-    # Compute damp from non-dead columns only, then add to all diagonals
-    damp = damp_factor * (torch.mean(diag[~dead]) if not dead.all() else torch.tensor(1.0, device=H.device)) if dead.any() else damp_factor * torch.mean(diag)
-    diag_idx = torch.arange(cols, device=H.device)
-    H[diag_idx, diag_idx] += damp
-    # Reset dead diag to just damp (smallest value) so they sort LAST in descending actorder
-    H[dead, dead] = damp
-    perm = torch.argsort(torch.diag(H), descending=(col_order == "desc"))
-    inv_perm = torch.argsort(perm)
-    W = t32[:, perm].clone()
-    W[:, dead[perm]] = 0
-    H = H[perm][:, perm]
-    try:
-        Hinv = torch.linalg.cholesky(H)
-        Hinv = torch.cholesky_inverse(Hinv)
-        Hinv = torch.linalg.cholesky(Hinv, upper=True)
-    except torch.linalg.LinAlgError:
-        return _quantize_int6_percentile(t32, clip_range)
-    # Single-pass mode: pre-compute scales once, run GPTQ once
-    if single_pass:
-        best_s = _precompute_row_scales(W, clip_range)
-        sf = best_s.float()
-        Q = _gptq_block_sweep(W, Hinv, sf, -clip_range, clip_range, block_size)
-        Q = Q[:, inv_perm]
-        return Q.cpu(), best_s.to(torch.float16).cpu()
-    # Multi-pass: search 5 percentiles, run GPTQ per percentile, pick best by MSE
-    best_q = None; best_scale = None; best_err = float('inf')
-    for pct in [0.9990, 0.9995, 0.9999, 0.99999, 1.0]:
-        if pct < 1.0:
-            row_clip = torch.quantile(t32.abs(), pct, dim=1)
-        else:
-            row_clip = t32.abs().amax(dim=1)
-        s = (row_clip / clip_range).clamp_min(1.0 / clip_range).to(torch.float16)
-        sf = s.float()
-        Q = _gptq_block_sweep(W, Hinv, sf, -clip_range, clip_range, block_size)
-        recon = Q.float() * sf[:, None]
-        mse = (W - recon).pow(2).mean().item()
-        if mse < best_err:
-            best_q, best_scale, best_err = Q, s, mse
-    best_q = best_q[:, inv_perm]
-    return best_q.cpu(), best_scale.cpu()
-
-def _quantize_int6_percentile(t32, clip_range=31):
-    """Fallback: percentile search (for 1D or no-Hessian cases)."""
-    if t32.ndim == 2:
-        best_q, best_s, best_err = None, None, float('inf')
-        for pct in [0.9990, 0.9995, 0.9999, 0.99999, 1.0]:
-            if pct < 1.0:
-                row_clip = torch.quantile(t32.abs(), pct, dim=1)
-            else:
-                row_clip = t32.abs().amax(dim=1)
-            s = (row_clip / clip_range).clamp_min(1.0 / clip_range).to(torch.float16)
-            q = torch.clamp(torch.round(t32 / s.float()[:, None]), -clip_range, clip_range).to(torch.int8)
-            recon = q.float() * s.float()[:, None]
-            err = (t32 - recon).pow(2).mean().item()
-            if err < best_err:
-                best_q, best_s, best_err = q, s, err
-        return best_q, best_s
-    amax = t32.abs().max().item()
-    scale = torch.tensor(amax / clip_range if amax > 0 else 1.0, dtype=torch.float16)
-    q = torch.clamp(torch.round(t32 / scale.float()), -clip_range, clip_range).to(torch.int8)
-    return q, scale
-
-def _unbank_state_dict(sd: dict[str, Tensor], num_layers: int) -> dict[str, Tensor]:
-    """Convert 3D bank tensors into individual 2D tensors with standard names."""
-    out: dict[str, Tensor] = {}
-    n = num_layers
-    for name, tensor in sd.items():
-        if name == "qo_bank":
-            for i in range(n):
-                out[f"blocks.{i}.attn.c_q.weight"] = tensor[i]
-                out[f"blocks.{i}.attn.proj.weight"] = tensor[n + i]
-        elif name == "kv_bank":
-            for i in range(n):
-                out[f"blocks.{i}.attn.c_k.weight"] = tensor[i]
-                out[f"blocks.{i}.attn.c_v.weight"] = tensor[n + i]
-        elif name == "mlp_up_bank":
-            for i in range(n):
-                out[f"blocks.{i}.mlp.fc.weight"] = tensor[i]
-        elif name == "mlp_down_bank":
-            for i in range(n):
-                out[f"blocks.{i}.mlp.proj.weight"] = tensor[i]
-        else:
-            out[name] = tensor
-    return out
-
-def _rebank_state_dict(sd: dict[str, Tensor], num_layers: int, template_sd: dict[str, Tensor] | None = None) -> dict[str, Tensor]:
-    """Convert individual 2D tensors back into 3D bank tensors."""
-    out: dict[str, Tensor] = {}
-    n = num_layers
-    consumed: set[str] = set()
-    qo_slices: list[Tensor | None] = [None] * (2 * n)
-    kv_slices: list[Tensor | None] = [None] * (2 * n)
-    up_slices: list[Tensor | None] = [None] * n
-    down_slices: list[Tensor | None] = [None] * n
-    _BANK_MAP = {
-        "attn.c_q.weight": (qo_slices, 0),
-        "attn.proj.weight": (qo_slices, n),
-        "attn.c_k.weight": (kv_slices, 0),
-        "attn.c_v.weight": (kv_slices, n),
-        "mlp.fc.weight": (up_slices, 0),
-        "mlp.proj.weight": (down_slices, 0),
-    }
-    for i in range(n):
-        for suffix, (target_list, offset) in _BANK_MAP.items():
-            key = f"blocks.{i}.{suffix}"
-            if key in sd:
-                target_list[offset + i] = sd[key]
-                consumed.add(key)
-    # Stack into bank tensors — validate all slices are present
-    for bank_name, slices in [("qo_bank", qo_slices), ("kv_bank", kv_slices), ("mlp_up_bank", up_slices), ("mlp_down_bank", down_slices)]:
-        if not any(s is not None for s in slices):
-            continue
-        missing = [i for i, s in enumerate(slices) if s is None]
-        if missing:
-            raise ValueError(f"_rebank_state_dict: {bank_name} missing slice indices {missing}")
-        out[bank_name] = torch.stack(slices)  # type: ignore[arg-type]
-    # Pass through non-banked params
-    for key, val in sd.items():
-        if key not in consumed:
-            out[key] = val
-    return out
-
-# --- Non-banked model for Hessian collection ---
-# This mirrors the unbanked state dict keys: blocks.{i}.attn.c_q/c_k/c_v/proj, blocks.{i}.mlp.fc/proj
-
-class _HessianAttn(nn.Module):
-    """Non-banked attention with CastedLinear layers for Hessian hooks."""
-    def __init__(self, dim, num_heads, num_kv_heads, rope_base, qk_gain_init):
-        super().__init__()
-        self.num_heads, self.num_kv_heads = num_heads, num_kv_heads
-        self.head_dim = dim // num_heads
-        kv_dim = num_kv_heads * self.head_dim
-        self.c_q = CastedLinear(dim, dim, bias=False)
-        self.c_k = CastedLinear(dim, kv_dim, bias=False)
-        self.c_v = CastedLinear(dim, kv_dim, bias=False)
-        self.proj = CastedLinear(dim, dim, bias=False)
-        self.q_gain = nn.Parameter(torch.full((num_heads,), qk_gain_init, dtype=torch.float32))
-        self.rope_dims = 0
-        self.rotary = Rotary(self.head_dim, base=rope_base, train_seq_len=1024)
-        self.use_xsa = False
-    def _xsa_efficient(self, y, v):
-        B, T, H, D = y.shape; Hkv = v.size(-2); group = H // Hkv
-        y_g = y.reshape(B, T, Hkv, group, D)
-        vn = F.normalize(v, dim=-1).unsqueeze(-2)
-        proj = (y_g * vn).sum(dim=-1, keepdim=True) * vn
-        return (y_g - proj).reshape(B, T, H, D)
-    def forward(self, x, v_embed=None):
-        bsz, seqlen, dim = x.shape
-        q = self.c_q(x).reshape(bsz, seqlen, self.num_heads, self.head_dim)
-        k = self.c_k(x).reshape(bsz, seqlen, self.num_kv_heads, self.head_dim)
-        v = self.c_v(x)
-        if v_embed is not None:
-            v = v + v_embed
-        v = v.reshape(bsz, seqlen, self.num_kv_heads, self.head_dim)
-        q = F.rms_norm(q, (q.size(-1),))
-        k = F.rms_norm(k, (k.size(-1),))
-        cos, sin = self.rotary(seqlen, x.device, q.dtype)
-        q = apply_rotary_emb(q, cos, sin, self.rope_dims)
-        k = apply_rotary_emb(k, cos, sin, self.rope_dims)
-        q = q * self.q_gain.to(dtype=q.dtype)[None, None, :, None]
-        y = F.scaled_dot_product_attention(
-            q.transpose(1, 2), k.transpose(1, 2), v.transpose(1, 2),
-            is_causal=True, enable_gqa=(self.num_kv_heads != self.num_heads),
-        ).transpose(1, 2)
-        if self.use_xsa:
-            y = self._xsa_efficient(y, v)
-        return self.proj(y.reshape(bsz, seqlen, dim))
-
-class _HessianMLP(nn.Module):
-    """Non-banked MLP with CastedLinear layers for Hessian hooks."""
-    def __init__(self, dim, mlp_mult):
-        super().__init__()
-        self.fc = CastedLinear(dim, int(mlp_mult * dim), bias=False)
-        self.proj = CastedLinear(int(mlp_mult * dim), dim, bias=False)
-    def forward(self, x):
-        return self.proj(F.leaky_relu(self.fc(x), negative_slope=0.3).square())
-
-class _HessianBlock(nn.Module):
-    def __init__(self, dim, num_heads, num_kv_heads, mlp_mult, rope_base, qk_gain_init, layer_idx=0, ln_scale=False):
-        super().__init__()
-        self.attn_norm = RMSNorm()
-        self.mlp_norm = RMSNorm()
-        self.attn = _HessianAttn(dim, num_heads, num_kv_heads, rope_base, qk_gain_init)
-        self.mlp = _HessianMLP(dim, mlp_mult)
-        self.attn_scale = nn.Parameter(torch.ones(dim, dtype=torch.float32))
-        self.mlp_scale = nn.Parameter(torch.ones(dim, dtype=torch.float32))
-        self.resid_mix = nn.Parameter(torch.stack((torch.ones(dim), torch.zeros(dim))).float())
-        self.ln_scale_factor = 1.0 / math.sqrt(layer_idx + 1) if ln_scale else 1.0
-    def forward(self, x, x0, v_embed=None):
-        mix = self.resid_mix.to(dtype=x.dtype)
-        x_in = mix[0][None, None, :] * x + mix[1][None, None, :] * x0
-        attn_out = self.attn(self.attn_norm(x_in) * self.ln_scale_factor, v_embed=v_embed)
-        x_out = x_in + self.attn_scale.to(dtype=x_in.dtype)[None, None, :] * attn_out
-        x_out = x_out + self.mlp_scale.to(dtype=x_out.dtype)[None, None, :] * self.mlp(self.mlp_norm(x_out) * self.ln_scale_factor)
-        return x_out
-
-class _HessianGPT(nn.Module):
-    """Non-banked GPT model matching unbanked state dict keys for Hessian collection."""
-    def __init__(self, vocab_size, num_layers, model_dim, num_heads, num_kv_heads,
-                 mlp_mult, tie_embeddings, logit_softcap, rope_base, qk_gain_init,
-                 ngram_buckets=0, ngram_heads=2, ngram_orders=2, ngram_dim_per_head=32,
-                 xsa_last_n=0, rope_dims=0, ln_scale=False,
-                 ve_enabled=False, ve_dim=128, ve_layers="9,10"):
-        super().__init__()
-        self.tie_embeddings = tie_embeddings
-        self.logit_softcap = logit_softcap
-        self.num_layers = num_layers
-        self.tok_emb = nn.Embedding(vocab_size, model_dim)
-        self.bigram = EngramLite(ngram_buckets, ngram_heads, ngram_orders, ngram_dim_per_head, model_dim) if ngram_buckets > 0 else None
-        self.smear = SmearGate(model_dim)
-        self.num_encoder_layers = num_layers // 2
-        self.num_decoder_layers = num_layers - self.num_encoder_layers
-        self.num_skip_weights = min(self.num_encoder_layers, self.num_decoder_layers)
-        self.skip_weights = nn.Parameter(torch.ones(self.num_skip_weights, model_dim, dtype=torch.float32))
-        self.skip_gates = nn.Parameter(torch.zeros(self.num_skip_weights, model_dim, dtype=torch.float32))
-        self.blocks = nn.ModuleList([
-            _HessianBlock(model_dim, num_heads, num_kv_heads, mlp_mult, rope_base, qk_gain_init,
-                          layer_idx=i, ln_scale=ln_scale)
-            for i in range(num_layers)
-        ])
-        if rope_dims > 0:
-            head_dim = model_dim // num_heads
-            for block in self.blocks:
-                block.attn.rope_dims = rope_dims
-                block.attn.rotary = Rotary(head_dim, base=rope_base, train_seq_len=1024, rope_dims=rope_dims)
-        if xsa_last_n > 0:
-            for i in range(max(0, num_layers - xsa_last_n), num_layers):
-                self.blocks[i].attn.use_xsa = True
-        kv_dim = num_kv_heads * (model_dim // num_heads)
-        self.ve_layer_indices = [int(x) for x in ve_layers.split(",") if x.strip()] if ve_enabled else []
-        if self.ve_layer_indices:
-            self.ve_shared = ValueEmbedding(vocab_size, ve_dim, kv_dim)
-            self.ve_layer_scales = nn.ParameterList([nn.Parameter(torch.ones(1, dtype=torch.float32)) for _ in self.ve_layer_indices])
-        else:
-            self.ve_shared = None
-            self.ve_layer_scales = nn.ParameterList()
-        self.final_norm = RMSNorm()
-        self.lm_head = None if tie_embeddings else CastedLinear(model_dim, vocab_size, bias=False)
-    def _get_ve(self, layer_idx, input_ids, ve_cache):
-        if self.ve_shared is None or layer_idx not in self.ve_layer_indices:
-            return None
-        if 've' not in ve_cache:
-            ve_cache['ve'] = self.ve_shared(input_ids)
-        ve_idx = self.ve_layer_indices.index(layer_idx)
-        return ve_cache['ve'] * self.ve_layer_scales[ve_idx].to(dtype=ve_cache['ve'].dtype)
-    def forward(self, input_ids, target_ids):
-        x = self.tok_emb(input_ids)
-        if self.bigram is not None:
-            x = x + self.bigram(input_ids)
-        x = F.rms_norm(x, (x.size(-1),))
-        x = self.smear(x)
-        x0 = x
-        skips = []
-        ve_cache = {}
-        for i in range(self.num_encoder_layers):
-            ve = self._get_ve(i, input_ids, ve_cache)
-            x = self.blocks[i](x, x0, v_embed=ve)
-            skips.append(x)
-        for i in range(self.num_decoder_layers):
-            bi = self.num_encoder_layers + i
-            if skips:
-                skip = skips.pop()
-                g = torch.sigmoid(self.skip_gates[i].to(dtype=x.dtype))[None, None, :]
-                scaled_skip = self.skip_weights[i].to(dtype=x.dtype)[None, None, :] * skip
-                x = torch.lerp(scaled_skip, x, g)
-            ve = self._get_ve(bi, input_ids, ve_cache)
-            x = self.blocks[bi](x, x0, v_embed=ve)
-        x = self.final_norm(x)
-        x_flat = x.reshape(-1, x.size(-1))
-        targets = target_ids.reshape(-1)
-        logits_proj = F.linear(x_flat, self.tok_emb.weight) if self.tie_embeddings else self.lm_head(x_flat)
-        logits = self.logit_softcap * torch.tanh(logits_proj / self.logit_softcap)
-        return F.cross_entropy(logits.float(), targets, reduction="mean")
-
-def collect_hessians(hessian_model, train_loader, args, device, grad_accum_steps, num_batches=256):
-    """Run calibration batches through a non-banked model, collecting H = X^T X for each CastedLinear."""
-    hessians = {}
-    hooks = []
-    for name, module in hessian_model.named_modules():
-        if isinstance(module, CastedLinear):
-            param_name = name + ".weight"
-            cols = module.weight.shape[1]
-            hessians[param_name] = torch.zeros(cols, cols, dtype=torch.float32, device=device)
-            def make_hook(pname):
-                def hook_fn(mod, inp, out):
-                    x = inp[0].detach()
-                    if x.ndim == 3:
-                        x = x.reshape(-1, x.shape[-1])
-                    hessians[pname] += (x.T @ x).float()  # bf16 matmul, fp32 accumulate
-                return hook_fn
-            h = module.register_forward_hook(make_hook(param_name))
-            hooks.append(h)
-    hessian_model.eval()
-    with torch.inference_mode(), torch.autocast(device_type="cuda", dtype=torch.bfloat16):
-        for _ in range(num_batches):
-            x, y = train_loader.next_batch(args.train_batch_tokens, args.train_seq_len, grad_accum_steps)
-            hessian_model(x, y)
-    for h in hooks:
-        h.remove()
-    ws = dist.get_world_size() if dist.is_available() and dist.is_initialized() else 1
-    for name in hessians:
-        H = hessians[name]
-        if ws > 1:
-            dist.all_reduce(H, op=dist.ReduceOp.SUM)
-        H /= (num_batches * ws)
-        hessians[name] = H.cpu()
-    hessian_model.train()
-    return hessians
-
-# --- Mixed precision bit allocation ---
-
-def _bits_to_range(bits: int) -> tuple[int, int]:
-    """Convert bit-width to (qmin, qmax)."""
-    return -(1 << (bits - 1)), (1 << (bits - 1)) - 1
-
-# Dynamic mixed-precision constants (tune after observing estimate vs actual)
-_MP_BYTES_PER_PARAM_INT5 = 0.46   # estimated compressed bytes per quantized param at int5
-_MP_COST_PER_EXTRA_BIT = 0.24     # additional compressed bytes per param per extra bit above int5
-_MP_NON_WEIGHT_COMPRESS = 0.55    # compression ratio for non-quantized tensors (fp16 embeds, scales)
-_MP_PRUNE_HEADROOM_FRAC = 0.02    # reserve 2% of byte budget so selective pruning only trims a small %
-
-
-def _allocate_bits_mixed(
-    hessian_map: dict[str, Tensor],
-    state_dict: dict[str, Tensor],
-    target_bytes: int = 16_000_000,
-    code_bytes: int = 0,
-) -> tuple[dict[str, int], list[tuple[str, int, float]], dict[str, float]]:
-    """Dynamically allocate int5-int7 per tensor group based on Hessian sensitivity
-    and compressed artifact size budget.
-
-    Greedy: promotes most-sensitive groups first (top group -> int7, rest -> int6)
-    until estimated compressed size approaches target_bytes minus pruning headroom.
-    Returns (tensor_name -> bits, [(group_key, bits, sensitivity)], estimate_info)."""
-    # 1. Group tensors by (layer, type) -- compute sensitivity and numel per group
-    group_traces: dict[str, list[float]] = {}
-    group_numel: dict[str, int] = {}
-    tensor_to_group: dict[str, str] = {}
-    for name, H in hessian_map.items():
-        trace_val = float(torch.trace(H).item()) / H.shape[0]
-        if not name.startswith("blocks."):
-            continue
-        dot2 = name.index(".", 7)
-        layer_idx = int(name[7:dot2])
-        gtype = "attn" if ".attn." in name else "mlp" if ".mlp." in name else "other"
-        gkey = f"layer.{layer_idx}.{gtype}"
-        group_traces.setdefault(gkey, []).append(trace_val)
-        tensor_to_group[name] = gkey
-        w = state_dict.get(name)
-        if w is not None:
-            group_numel[gkey] = group_numel.get(gkey, 0) + w.numel()
-    group_sensitivity = {k: sum(v) / len(v) for k, v in group_traces.items()}
-    ranked = sorted(group_sensitivity.items(), key=lambda x: x[1], reverse=True)
-
-    # 2. Estimate baseline compressed size (all quantized weights at int5)
-    total_quant_numel = sum(group_numel.values())
-    non_weight_raw = sum(
-        t.numel() * t.element_size() for name, t in state_dict.items()
-        if name not in hessian_map
-    )
-    base_estimate = (
-        code_bytes
-        + int(non_weight_raw * _MP_NON_WEIGHT_COMPRESS)
-        + int(total_quant_numel * _MP_BYTES_PER_PARAM_INT5)
-    )
-    budget = int(target_bytes * (1.0 - _MP_PRUNE_HEADROOM_FRAC)) - base_estimate
-
-    # Early exit: if base estimate already exceeds budget, return all int5
-    if budget <= 0:
-        bit_allocation = {tname: 5 for tname in tensor_to_group}
-        log_entries = [(gkey, 5, group_sensitivity[gkey]) for gkey, _ in ranked]
-        estimate_info = {
-            "base_mb": base_estimate / 1e6, "promoted_mb": 0.0,
-            "total_mb": base_estimate / 1e6, "budget_mb": target_bytes / 1e6,
-            "headroom_kb": 0.0, "prune_room_bytes": target_bytes - base_estimate,
-            "warning": "budget_exhausted",
-        }
-        return bit_allocation, log_entries, estimate_info
-
-    # 3. Greedy promotion: most sensitive first
-    #    Pass 1: try int7 for top group. Pass 2: fill remaining with int6.
-    group_bits: dict[str, int] = {gkey: 5 for gkey, _ in ranked}
-    estimated_extra = 0
-    if ranked:
-        top_gkey = ranked[0][0]
-        top_numel = group_numel.get(top_gkey, 0)
-        cost_int7 = int(top_numel * _MP_COST_PER_EXTRA_BIT * 2)
-        cost_int6 = int(top_numel * _MP_COST_PER_EXTRA_BIT * 1)
-        if top_numel > 0 and cost_int7 <= budget:
-            group_bits[top_gkey] = 7
-            estimated_extra += cost_int7
-        elif top_numel > 0 and cost_int6 <= budget:
-            group_bits[top_gkey] = 6  # int7 doesn't fit, fall back to int6
-            estimated_extra += cost_int6
-    for gkey, _sens in ranked:
-        if group_bits[gkey] > 5:
-            continue  # already promoted
-        numel = group_numel.get(gkey, 0)
-        if numel == 0:
-            continue
-        cost = int(numel * _MP_COST_PER_EXTRA_BIT)  # 1 extra bit for int6
-        if estimated_extra + cost <= budget:
-            group_bits[gkey] = 6
-            estimated_extra += cost
-        # don't break -- smaller groups later in the list may still fit
-
-    # 4. Map back to tensor names
-    bit_allocation: dict[str, int] = {}
-    for tname, gkey in tensor_to_group.items():
-        bit_allocation[tname] = group_bits[gkey]
-    log_entries = [(gkey, group_bits[gkey], group_sensitivity[gkey]) for gkey, _ in ranked]
-
-    # 5. Size estimate metadata for caller to log via log0()
-    est_total = base_estimate + estimated_extra
-    headroom_bytes = int(target_bytes * _MP_PRUNE_HEADROOM_FRAC)
-    estimate_info = {
-        "base_mb": base_estimate / 1e6,
-        "promoted_mb": estimated_extra / 1e6,
-        "total_mb": est_total / 1e6,
-        "budget_mb": target_bytes / 1e6,
-        "headroom_kb": headroom_bytes / 1e3,
-        "prune_room_bytes": target_bytes - est_total - headroom_bytes,
-    }
-    return bit_allocation, log_entries, estimate_info
-
-
-def mixed_quantize_int6(state_dict: dict[str, Tensor], int6_cats: set[str], hessians: dict[str, Tensor] | None = None,
-                        bit_allocation: dict[str, int] | None = None, gptq_damp: float = 0.01,
-                        block_size: int = 128, col_order: str = "desc", single_pass: bool = False):
-    result: dict[str, Tensor] = {}
-    meta: dict[str, object] = {}
-    for name, tensor in state_dict.items():
-        t = tensor.detach().cpu().contiguous()
-        cat = _classify_param(name)
-        if not t.is_floating_point() or t.numel() <= 65536 or name == "tok_emb.weight":
-            result[name] = t.to(torch.float16) if t.is_floating_point() else t
-            meta[name] = "passthrough"
-            continue
-        if any(p in name for p in CONTROL_TENSOR_NAME_PATTERNS):
-            result[name] = t.float()
-            meta[name] = "passthrough_ctrl"
-            continue
-        if cat in int6_cats and t.ndim >= 1:
-            # Use mixed precision bit allocation if available
-            bits = bit_allocation.get(name, 6) if bit_allocation else 6
-            qmin, qmax = _bits_to_range(bits)
-            cr = qmax
-            H = hessians.get(name) if hessians else None
-            if H is not None:
-                q, s = quantize_int6_gptq(t, hessian=H, clip_range=cr, block_size=block_size,
-                                           damp_factor=gptq_damp, col_order=col_order, single_pass=single_pass)
-            else:
-                q, s = quantize_int6_per_row(t, clip_range=cr)
-            result[name + ".q"] = q
-            result[name + ".scale"] = s
-            meta[name] = {"type": f"int{bits}"}
-        else:
-            q, s = quantize_float_tensor(t)
-            result[name + ".q"] = q
-            result[name + ".scale"] = s
-            meta[name] = {"type": "int8"}
-    return result, meta
-def dequantize_mixed_int6(result: dict[str, Tensor], meta: dict[str, object],
-                          template_sd: dict[str, Tensor]) -> dict[str, Tensor]:
-    out: dict[str, Tensor] = {}
-    for name, orig in template_sd.items():
-        info = meta.get(name)
-        if info is None:
-            continue
-        orig_dtype = orig.dtype
-        if info in ("passthrough", "passthrough_ctrl", "passthrough_fp16"):
-            t = result[name]
-            if t.dtype == torch.float16 and orig_dtype in (torch.float32, torch.bfloat16):
-                t = t.to(orig_dtype)
-            out[name] = t
-            continue
-        q, s = result[name + ".q"], result[name + ".scale"]
-        if s.ndim > 0:
-            out[name] = (q.float() * s.float().view(q.shape[0], *([1] * (q.ndim - 1)))).to(orig_dtype)
-        else:
-            out[name] = (q.float() * float(s.item())).to(orig_dtype)
-    return out
-
-# --- Training ---
-
-def main() -> None:
-    code = Path(__file__).read_text(encoding="utf-8")
-    args = Hyperparameters()
-    rank = int(os.environ.get("RANK", "0"))
-    world_size = int(os.environ.get("WORLD_SIZE", "1"))
-    local_rank = int(os.environ.get("LOCAL_RANK", "0"))
-    distributed = "RANK" in os.environ and "WORLD_SIZE" in os.environ and world_size > 1
-    if world_size <= 0:
-        raise ValueError(f"WORLD_SIZE must be positive, got {world_size}")
-    if 8 % world_size != 0:
-        raise ValueError(f"WORLD_SIZE={world_size} must divide 8 so grad_accum_steps stays integral")
-    grad_accum_steps = 8 // world_size
-    grad_scale = 1.0 / grad_accum_steps
-    if not torch.cuda.is_available():
-        raise RuntimeError("CUDA is required")
-    device = torch.device("cuda", local_rank)
-    torch.cuda.set_device(device)
-    if distributed:
-        dist.init_process_group(backend="nccl", device_id=device)
-        dist.barrier()
-    master_process = rank == 0
-    torch.backends.cuda.matmul.allow_tf32 = True
-    torch.backends.cudnn.allow_tf32 = True
-    from torch.backends.cuda import enable_cudnn_sdp, enable_flash_sdp, enable_math_sdp, enable_mem_efficient_sdp
-    enable_cudnn_sdp(True)
-    enable_flash_sdp(True)
-    enable_mem_efficient_sdp(True)
-    enable_math_sdp(True)
-    logfile = None
-    if master_process:
-        os.makedirs("logs", exist_ok=True)
-        logfile = f"logs/{args.run_id}.txt"
-        print(logfile)
-    def log0(msg: str, console: bool = True) -> None:
-        if not master_process:
-            return
-        if console:
-            print(msg)
-        if logfile is not None:
-            with open(logfile, "a", encoding="utf-8") as f:
-                print(msg, file=f)
-    log0(code, console=False)
-    log0("=" * 100, console=False)
-    log0(f"Running Python {sys.version}", console=False)
-    log0(f"Running PyTorch {torch.__version__}", console=False)
-    log0(
-        subprocess.run(["nvidia-smi"], stdout=subprocess.PIPE, stderr=subprocess.PIPE, text=True, check=False).stdout,
-        console=False,
-    )
-    log0("=" * 100, console=False)
-    random.seed(args.seed)
-    np.random.seed(args.seed)
-    torch.manual_seed(args.seed)
-    torch.cuda.manual_seed_all(args.seed)
-    if not args.tokenizer_path.endswith(".model"):
-        raise ValueError(f"Script only setup for SentencePiece .model file: {args.tokenizer_path}")
-    sp = spm.SentencePieceProcessor(model_file=args.tokenizer_path)
-    if int(sp.vocab_size()) != args.vocab_size:
-        raise ValueError(
-            f"VOCAB_SIZE={args.vocab_size} does not match tokenizer vocab_size={int(sp.vocab_size())}"
-        )
-    dataset_dir = Path(args.data_path).resolve()
-    actual_train_files = len(list(dataset_dir.glob("fineweb_train_*.bin")))
-    effective_eval_seq_len = args.eval_seq_len if args.eval_seq_len > 0 else args.train_seq_len
-    val_seq_len = max(args.train_seq_len, effective_eval_seq_len)
-    val_tokens = load_validation_tokens(args.val_files, val_seq_len)
-    base_bytes_lut, has_leading_space_lut, is_boundary_token_lut = build_sentencepiece_luts(
-        sp, args.vocab_size, device
-    )
-    log0(f"val_bpb:enabled tokenizer_kind=sentencepiece tokenizer_path={args.tokenizer_path}")
-    log0(f"train_loader:dataset:{dataset_dir.name} train_shards:{actual_train_files}")
-    log0(f"val_loader:shards pattern={args.val_files} tokens:{val_tokens.numel() - 1}")
-    if not args.load_snapshot:
-        CastedLinear._qat_enabled = False  # late_qat enables mid-run when LR scale drops
-        CastedLinear._qat_soft_round = False
-        CastedLinear._qat_clip_pct = args.qat_clip_pct
-        base_model = GPT(
-            vocab_size=args.vocab_size,
-            num_layers=args.num_layers,
-            model_dim=args.model_dim,
-            num_heads=args.num_heads,
-            num_kv_heads=args.num_kv_heads,
-            mlp_mult=args.mlp_mult,
-            tie_embeddings=args.tie_embeddings,
-            tied_embed_init_std=args.tied_embed_init_std,
-            logit_softcap=args.logit_softcap,
-            rope_base=args.rope_base,
-            qk_gain_init=args.qk_gain_init,
-            ngram_buckets=args.ngram_buckets,
-            ngram_heads=args.ngram_heads,
-            ngram_orders=args.ngram_orders,
-            ngram_dim_per_head=args.ngram_dim_per_head,
-            xsa_last_n=args.xsa_last_n,
-            rope_dims=args.rope_dims,
-            ln_scale=args.ln_scale,
-            ve_enabled=args.ve_enabled,
-            ve_dim=args.ve_dim,
-            ve_layers=args.ve_layers,
-        ).to(device).bfloat16()
-        # Banks stay FP32 (like CastedLinear weights), cast to BF16 in forward
-        base_model.qo_bank.data = base_model.qo_bank.data.float()
-        base_model.kv_bank.data = base_model.kv_bank.data.float()
-        base_model.mlp_up_bank.data = base_model.mlp_up_bank.data.float()
-        base_model.mlp_down_bank.data = base_model.mlp_down_bank.data.float()
-        for module in base_model.modules():
-            if isinstance(module, CastedLinear):
-                module.float()
-        restore_low_dim_params_to_fp32(base_model)
-        # No DDP -- Parallel Muon handles bank grad communication via reduce-scatter,
-        # and non-bank grads are manually all-reduced before Adam steps.
-        # Compile NS functions for ~2x speedup on the orthogonalization hot path
-        global zeropower_via_newtonschulz5, _post_ns_normalize
-        zeropower_via_newtonschulz5 = torch.compile(zeropower_via_newtonschulz5)
-        _post_ns_normalize = torch.compile(_post_ns_normalize)
-        compiled_model = torch.compile(base_model, dynamic=False, fullgraph=True)
-        model = compiled_model
-
-        # Optimizer split:
-        # - 4 parameter banks -> Muon (batched Newton-Schulz)
-        # - token embedding -> Adam (with embed_beta1)
-        # - scalars/control tensors -> Adam
-        # - EngramLite proj -> Muon (small 2D matrix)
-        matrix_params = [
-            base_model.qo_bank, base_model.kv_bank,
-            base_model.mlp_up_bank, base_model.mlp_down_bank,
-        ]
-        block_named_params = list(base_model.blocks.named_parameters())
-        scalar_params = [
-            p
-            for name, p in block_named_params
-            if p.ndim < 2 or any(pattern in name for pattern in CONTROL_TENSOR_NAME_PATTERNS)
-        ]
-        if base_model.skip_weights.numel() > 0:
-            scalar_params.append(base_model.skip_weights)
-        if hasattr(base_model, 'skip_gates') and base_model.skip_gates.numel() > 0:
-            scalar_params.append(base_model.skip_gates)
-        scalar_params.append(base_model.smear.gate)
-        if base_model.bigram is not None:
-            scalar_params.append(base_model.bigram.ngram_gate)
-        token_lr = args.tied_embed_lr if args.tie_embeddings else args.embed_lr
-        tok_params = [{"params": [base_model.tok_emb.weight], "lr": token_lr, "base_lr": token_lr}]
-        if base_model.bigram is not None:
-            tok_params.append({"params": [base_model.bigram.embed.weight], "lr": token_lr, "base_lr": token_lr})
-        ve_proj_weight = None
-        if base_model.ve_shared is not None:
-            tok_params.append({"params": [base_model.ve_shared.embed.weight], "lr": token_lr, "base_lr": token_lr})
-            if base_model.ve_shared.proj is not None:
-                ve_proj_weight = base_model.ve_shared.proj.weight  # routed to Muon below
-            scalar_params.append(base_model.ve_shared.scale)
-            for s in base_model.ve_layer_scales:
-                scalar_params.append(s)
-        optimizer_tok = torch.optim.AdamW(
-            tok_params,
-            betas=(args.embed_beta1, args.beta2),
-            eps=args.adam_eps,
-            weight_decay=args.adam_wd,
-            fused=True,
-        )
-        optimizer_muon = Muon(
-            matrix_params,
-            lr=args.matrix_lr,
-            momentum=args.muon_momentum,
-            backend_steps=args.muon_backend_steps,
-            weight_decay=args.muon_wd,
-            post_norm=args.muon_post_norm,
-        )
-        for group in optimizer_muon.param_groups:
-            group["base_lr"] = args.matrix_lr
-        optimizer_scalar = torch.optim.AdamW(
-            [{"params": scalar_params, "lr": args.scalar_lr, "base_lr": args.scalar_lr}],
-            betas=(args.beta1, args.beta2),
-            eps=args.adam_eps,
-            weight_decay=args.adam_wd,
-            fused=True,
-        )
-        # Non-bank params that need manual all-reduce (replicated across GPUs)
-        replicated_params = list(optimizer_tok.param_groups[0]["params"])
-        for pg in optimizer_tok.param_groups[1:]:
-            replicated_params.extend(pg["params"])
-        replicated_params.extend(scalar_params)
-
-        optimizers: list[torch.optim.Optimizer] = [optimizer_tok, optimizer_muon, optimizer_scalar]
-
-        # EngramLite proj -> Muon (small 2D matrix, not banked)
-        if base_model.bigram is not None and base_model.bigram.proj is not None:
-            optimizer_bigram_proj = Muon(
-                [base_model.bigram.proj.weight],
-                lr=args.matrix_lr,
-                momentum=args.muon_momentum,
-                backend_steps=args.muon_backend_steps,
-                weight_decay=args.muon_wd,
-                post_norm=args.muon_post_norm,
-            )
-            for group in optimizer_bigram_proj.param_groups:
-                group["base_lr"] = args.matrix_lr
-            optimizers.append(optimizer_bigram_proj)
-            replicated_params.append(base_model.bigram.proj.weight)
-
-        # VE proj -> Muon (2D matrix, benefits from Newton-Schulz)
-        if ve_proj_weight is not None:
-            optimizer_ve_proj = Muon(
-                [ve_proj_weight],
-                lr=args.matrix_lr,
-                momentum=args.muon_momentum,
-                backend_steps=args.muon_backend_steps,
-                weight_decay=args.muon_wd,
-                post_norm=args.muon_post_norm,
-            )
-            for group in optimizer_ve_proj.param_groups:
-                group["base_lr"] = args.matrix_lr
-            optimizers.append(optimizer_ve_proj)
-            replicated_params.append(ve_proj_weight)
-
-        optimizer_head = None
-        if base_model.lm_head is not None:
-            optimizer_head = torch.optim.Adam(
-                [{"params": [base_model.lm_head.weight], "lr": args.head_lr, "base_lr": args.head_lr}],
-                betas=(args.head_beta1, args.beta2),
-                eps=args.adam_eps,
-                fused=True,
-            )
-            replicated_params.append(base_model.lm_head.weight)
-            optimizers.append(optimizer_head)
-
-        # Pre-build flat buffer for coalesced all-reduce of non-bank grads (saves ~0.5-1ms/step on multi-GPU)
-        _nb_grad_numel = [p.numel() for p in replicated_params]
-        _nb_grad_buf = torch.zeros(sum(_nb_grad_numel), device=device, dtype=torch.float32) if distributed else None
-
-        n_params = sum(p.numel() for p in base_model.parameters())
-        log0(f"model_params:{n_params}")
-        xsa_layers = [i for i, b in enumerate(base_model.blocks) if b.attn.use_xsa]
-        log0(f"XSA:last_{args.xsa_last_n} active_layers:{xsa_layers}")
-        log0(f"world_size:{world_size} grad_accum_steps:{grad_accum_steps}")
-        log0("sdp_backends:cudnn=True flash=True mem_efficient=True math=True")
-        log0(f"attention_mode:gqa num_heads:{args.num_heads} num_kv_heads:{args.num_kv_heads}")
-        log0(
-            f"tie_embeddings:{args.tie_embeddings} embed_lr:{token_lr} "
-            f"head_lr:{args.head_lr if base_model.lm_head is not None else 0.0} "
-            f"matrix_lr:{args.matrix_lr} scalar_lr:{args.scalar_lr}"
-        )
-        log0(
-            f"train_batch_tokens:{args.train_batch_tokens} train_seq_len:{args.train_seq_len} "
-            f"iterations:{args.iterations} warmup_steps:{args.warmup_steps} "
-            f"max_wallclock_seconds:{args.max_wallclock_seconds:.3f}"
-        )
-        log0(f"muon:post_norm:{args.muon_post_norm} embed_beta1:{args.embed_beta1} head_beta1:{args.head_beta1}")
-        log0(f"seed:{args.seed}")
-        train_loader = DistributedTokenLoader(args.train_files, rank, world_size, device)
-        def zero_grad_all() -> None:
-            for opt in optimizers:
-                opt.zero_grad(set_to_none=True)
-        max_wallclock_ms = 1000.0 * args.max_wallclock_seconds if args.max_wallclock_seconds > 0 else None
-        # Reserve time for GPTQ calibration so LR warmdown, QAT, and wallclock cap all see the effective budget
-        if max_wallclock_ms is not None and args.gptq_calib_batches > 0 and args.gptq_reserve_ms > 0:
-            max_wallclock_ms -= args.gptq_reserve_ms
-            log0(f"gptq:reserving {args.gptq_reserve_ms:.0f}ms from training budget (effective cap: {max_wallclock_ms:.0f}ms)")
-        def lr_mul(step: int, elapsed_ms: float) -> float:
-            if args.warmdown_iters <= 0:
-                return 1.0
-            if max_wallclock_ms is None:
-                warmdown_start = max(args.iterations - args.warmdown_iters, 0)
-                return max((args.iterations - step) / max(args.warmdown_iters, 1), args.lr_floor) if warmdown_start <= step < args.iterations else 1.0
-            step_ms = elapsed_ms / max(step, 1)
-            warmdown_ms = min(args.warmdown_iters * step_ms, max_wallclock_ms)
-            remaining_ms = max(max_wallclock_ms - elapsed_ms, 0.0)
-            return max(remaining_ms / max(warmdown_ms, 1e-9), args.lr_floor) if remaining_ms <= warmdown_ms else 1.0
-        if args.warmup_steps > 0:
-            initial_model_state = {name: tensor.detach().cpu().clone() for name, tensor in base_model.state_dict().items()}
-            initial_optimizer_states = [copy.deepcopy(opt.state_dict()) for opt in optimizers]
-            model.train()
-            for warmup_step in range(args.warmup_steps):
-                zero_grad_all()
-                for micro_step in range(grad_accum_steps):
-                    x, y = train_loader.next_batch(args.train_batch_tokens, args.train_seq_len, grad_accum_steps)
-                    with torch.autocast(device_type="cuda", dtype=torch.bfloat16, enabled=True):
-                        warmup_loss = model(x, y)
-                    (warmup_loss * grad_scale).backward()
-                # All-reduce all grads for warmup (simple, not optimized)
-                if distributed:
-                    for p in base_model.parameters():
-                        if p.grad is not None:
-                            dist.all_reduce(p.grad, op=dist.ReduceOp.AVG)
-                for opt in optimizers:
-                    opt.step()
-                zero_grad_all()
-                if args.warmup_steps <= 20 or (warmup_step + 1) % 10 == 0 or warmup_step + 1 == args.warmup_steps:
-                    log0(f"warmup_step:{warmup_step + 1}/{args.warmup_steps}")
-            base_model.load_state_dict(initial_model_state, strict=True)
-            for opt, state in zip(optimizers, initial_optimizer_states, strict=True):
-                opt.load_state_dict(state)
-            # Reset Muon shard_mom buffers — not captured by state_dict()
-            if optimizer_muon._built:
-                for m in optimizer_muon._bank_meta:
-                    m['shard_mom'].zero_()
-            zero_grad_all()
-            train_loader = DistributedTokenLoader(args.train_files, rank, world_size, device)
-        swa_state: dict[str, Tensor] | None = None
-        swa_count = 0
-        ema_state: dict[str, Tensor] | None = None
-        _ema_pairs: list[tuple[Tensor, Tensor]] | None = None
-        if args.ema_enabled:
-            ema_state = {name: p.data.detach().float().clone() for name, p in base_model.named_parameters()}
-            _ema_pairs = [(ema_state[name], p) for name, p in base_model.named_parameters()]
-        training_time_ms = 0.0
-        _qat_start_step = 0
-        _qat_total_steps = 0
-        _train_loss = torch.zeros((), device=device)
-        _cap_tensor = torch.zeros(1, device=device, dtype=torch.int32) if distributed else None
-        stop_after_step: int | None = None
-        torch.cuda.synchronize()
-        t0 = time.perf_counter()
-        step = 0
-        while True:
-            last_step = step == args.iterations or (stop_after_step is not None and step >= stop_after_step)
-            if last_step:
-                torch.cuda.synchronize()
-                training_time_ms += 1000.0 * (time.perf_counter() - t0)
-                log0(
-                    f"step:{step}/{args.iterations} "
-                    f"train_time:{training_time_ms:.0f}ms step_avg:{training_time_ms / max(step, 1):.2f}ms"
-                )
-                if stop_after_step is not None and step < args.iterations:
-                    log0(
-                        f"stopping_early: wallclock_cap train_time:{training_time_ms:.0f}ms "
-                        f"step:{step}/{args.iterations}"
-                    )
-                break
-            elapsed_ms = training_time_ms + 1000.0 * (time.perf_counter() - t0)
-            scale = lr_mul(step, elapsed_ms)
-            if args.late_qat and args.qat_threshold > 0 and not CastedLinear._qat_enabled and scale < args.qat_threshold:
-                CastedLinear._qat_enabled = True
-                _qat_start_step = step
-                # Estimate remaining QAT steps using best available info
-                if stop_after_step is not None:
-                    _qat_total_steps = max(stop_after_step - step, 1)
-                elif max_wallclock_ms is not None:
-                    _step_ms = elapsed_ms / max(step, 1)
-                    _qat_total_steps = max(int((max_wallclock_ms - elapsed_ms) / max(_step_ms, 1e-9)), 1)
-                else:
-                    _qat_total_steps = max(args.iterations - step, 1)
-                if args.soft_round_qat:
-                    CastedLinear._qat_soft_round = True
-                    CastedLinear._qat_soft_alpha = torch.tensor(1.0, device=device)
-                log0(f"late_qat:enabled step:{step} scale:{scale:.4f} soft_round={args.soft_round_qat} est_steps:{_qat_total_steps}")
-            elif CastedLinear._qat_soft_round:
-                # Soft-round alpha ramp: 1→16 over QAT phase (tensor value, no torch.compile recompiles)
-                # Re-estimate if wallclock cap set after QAT started (tighter bound)
-                if stop_after_step is not None:
-                    _qat_total_steps = max(stop_after_step - _qat_start_step, 1)
-                qat_progress = min((step - _qat_start_step) / _qat_total_steps, 1.0)
-                with torch.no_grad():
-                    CastedLinear._qat_soft_alpha.fill_(1.0 + 15.0 * qat_progress)
-            zero_grad_all()
-            _train_loss.zero_()
-            for micro_step in range(grad_accum_steps):
-                x, y = train_loader.next_batch(args.train_batch_tokens, args.train_seq_len, grad_accum_steps)
-                with torch.autocast(device_type="cuda", dtype=torch.bfloat16, enabled=True):
-                    loss = model(x, y)
-                _train_loss += loss.detach()
-                (loss * grad_scale).backward()
-            _train_loss /= grad_accum_steps
-            frac = min(step / args.muon_momentum_warmup_steps, 1.0) if args.muon_momentum_warmup_steps > 0 else 1.0
-            muon_momentum = (1 - frac) * args.muon_momentum_warmup_start + frac * args.muon_momentum
-            for group in optimizer_muon.param_groups:
-                group["momentum"] = muon_momentum
-            for opt in optimizers:
-                for group in opt.param_groups:
-                    group["lr"] = group["base_lr"] * scale
-            if args.grad_clip_norm > 0:
-                torch.nn.utils.clip_grad_norm_(base_model.parameters(), args.grad_clip_norm)
-            # === 3-phase overlapped optimizer step ===
-            # Phase 1: Launch async reduce-scatter for banks (biggest first)
-            optimizer_muon.launch_reduce_scatters()
-            # Phase 2: All-reduce non-bank grads (coalesced) + step Adam (while bank RS is in-flight)
-            if distributed:
-                offset = 0
-                for p, n in zip(replicated_params, _nb_grad_numel):
-                    if p.grad is not None:
-                        _nb_grad_buf[offset:offset + n].copy_(p.grad.reshape(-1))
-                    else:
-                        _nb_grad_buf[offset:offset + n].zero_()
-                    offset += n
-                dist.all_reduce(_nb_grad_buf, op=dist.ReduceOp.AVG)
-                offset = 0
-                for p, n in zip(replicated_params, _nb_grad_numel):
-                    if p.grad is not None:
-                        p.grad.copy_(_nb_grad_buf[offset:offset + n].reshape_as(p.grad))
-                    offset += n
-            optimizer_tok.step()
-            optimizer_scalar.step()
-            # Step non-bank Muon optimizers (e.g., bigram proj)
-            for opt in optimizers:
-                if opt is not optimizer_muon and opt is not optimizer_tok and opt is not optimizer_scalar:
-                    if optimizer_head is not None and opt is optimizer_head:
-                        opt.step()
-                    elif isinstance(opt, Muon):
-                        opt.step()
-            # Phase 3: Wait for RS, local NS5, all-gather (banks processed last)
-            optimizer_muon.step()
-            zero_grad_all()
-            # EMA update (skip .float() for already-fp32 bank params)
-            if _ema_pairs is not None:
-                _ema_weight = 1.0 - args.ema_decay
-                with torch.no_grad():
-                    for ema_buf, p in _ema_pairs:
-                        ema_buf.lerp_(
-                            p.data if p.data.dtype == torch.float32 else p.data.float(),
-                            _ema_weight,
-                        )
-            step += 1
-            approx_training_time_ms = training_time_ms + 1000.0 * (time.perf_counter() - t0)
-            if args.swa_enabled and scale < args.swa_threshold and step % args.swa_every == 0:
-                if swa_state is None:
-                    swa_state = {name: p.data.detach().float().clone() for name, p in base_model.named_parameters()}
-                    swa_count = 1
-                    log0(f"swa:start step:{step}")
-                else:
-                    for name, p in base_model.named_parameters():
-                        swa_state[name].add_(p.data if p.data.dtype == torch.float32 else p.data.float())
-                    swa_count += 1
-            should_log_train = (
-                args.train_log_every > 0
-                and (step <= 10 or step % args.train_log_every == 0 or stop_after_step is not None)
-            )
-            if should_log_train:
-                log0(
-                    f"step:{step}/{args.iterations} train_loss:{_train_loss.item():.4f} "
-                    f"train_time:{approx_training_time_ms:.0f}ms step_avg:{approx_training_time_ms / step:.2f}ms"
-                )
-            reached_cap = max_wallclock_ms is not None and approx_training_time_ms >= max_wallclock_ms
-            if distributed and max_wallclock_ms is not None:
-                _cap_tensor.fill_(int(reached_cap))
-                dist.all_reduce(_cap_tensor, op=dist.ReduceOp.MAX)
-                reached_cap = bool(_cap_tensor.item())
-            if stop_after_step is None and reached_cap:
-                stop_after_step = step
-        log0(
-            f"peak memory allocated: {torch.cuda.max_memory_allocated() // 1024 // 1024} MiB "
-            f"reserved: {torch.cuda.max_memory_reserved() // 1024 // 1024} MiB"
-        )
-        # Apply weight averaging: SWA if available, else EMA
-        if args.swa_enabled and swa_state is not None and swa_count > 1:
-            log0(f"swa:applying averaged {swa_count} checkpoints (source=raw)")
-            with torch.no_grad():
-                for name, p in base_model.named_parameters():
-                    if name in swa_state:
-                        p.data.copy_((swa_state[name] / swa_count).to(dtype=p.dtype))
-            del swa_state
-        elif ema_state is not None:
-            log0("ema:applying EMA weights")
-            with torch.no_grad():
-                for name, p in base_model.named_parameters():
-                    if name in ema_state:
-                        p.data.copy_(ema_state[name].to(dtype=p.dtype))
-            del ema_state
-        full_state_dict = base_model.state_dict()
-        export_sd = full_state_dict
-        if master_process:
-            torch.save(export_sd, "final_model.pt")
-            model_bytes = os.path.getsize("final_model.pt")
-            code_bytes = len(code.encode("utf-8"))
-            log0(f"Serialized model: {model_bytes} bytes")
-            log0(f"Code size: {code_bytes} bytes")
-        # Unbank 3D tensors into individual 2D tensors for quantization
-        sd_cpu = {k: v.detach().cpu() for k, v in export_sd.items()}
-        unbanked_sd = _unbank_state_dict(sd_cpu, args.num_layers)
-
-        # Full GPTQ: collect Hessians via a temporary non-banked model
-        _gptq_t0 = time.perf_counter()
-        log0(f"gptq:building non-banked model for Hessian collection...")
-        hessian_model = _HessianGPT(
-            vocab_size=args.vocab_size, num_layers=args.num_layers, model_dim=args.model_dim,
-            num_heads=args.num_heads, num_kv_heads=args.num_kv_heads, mlp_mult=args.mlp_mult,
-            tie_embeddings=args.tie_embeddings, logit_softcap=args.logit_softcap,
-            rope_base=args.rope_base, qk_gain_init=args.qk_gain_init,
-            ngram_buckets=args.ngram_buckets, ngram_heads=args.ngram_heads,
-            ngram_orders=args.ngram_orders, ngram_dim_per_head=args.ngram_dim_per_head,
-            xsa_last_n=args.xsa_last_n, rope_dims=args.rope_dims, ln_scale=args.ln_scale,
-            ve_enabled=args.ve_enabled, ve_dim=args.ve_dim, ve_layers=args.ve_layers,
-        ).to(device).bfloat16()
-        for m in hessian_model.modules():
-            if isinstance(m, CastedLinear):
-                m.float()
-        restore_low_dim_params_to_fp32(hessian_model)
-        # Load unbanked weights into the non-banked model
-        hessian_model.load_state_dict(
-            {k: v.to(device) for k, v in unbanked_sd.items() if k in hessian_model.state_dict()},
-            strict=False,
-        )
-        log0(f"gptq:calibrating with {args.gptq_calib_batches} batches...")
-        calib_loader = DistributedTokenLoader(args.train_files, rank, world_size, device)
-        hessians = collect_hessians(hessian_model, calib_loader, args, device, grad_accum_steps,
-                                    num_batches=args.gptq_calib_batches)
-        log0(f"gptq:collected hessians for {len(hessians)} layers")
-        _gptq_ms = 1000.0 * (time.perf_counter() - _gptq_t0)
-        _full_budget_ms = (max_wallclock_ms + args.gptq_reserve_ms) if max_wallclock_ms is not None else 0
-        log0(f"gptq:budget_check train:{approx_training_time_ms:.0f}ms + gptq:{_gptq_ms:.0f}ms = {approx_training_time_ms + _gptq_ms:.0f}ms (budget:{_full_budget_ms:.0f}ms)")
-        del hessian_model
-        torch.cuda.empty_cache()
-
-        # Snapshot: save unbanked_sd + hessians + sd_cpu so compression can be re-run without retraining
-        if args.snapshot_post_hessian:
-            if master_process:
-                snap_path = os.environ.get("SNAPSHOT_PATH", "snapshot_post_hessian.pt")
-                log0(f"snapshot:saving to {snap_path}...")
-                torch.save({"unbanked_sd": unbanked_sd, "hessians": {k: v.cpu() for k, v in hessians.items()}, "sd_cpu": sd_cpu}, snap_path)
-                snap_mb = os.path.getsize(snap_path) / 1e6
-                log0(f"snapshot:saved {snap_mb:.1f}MB — exiting (use LOAD_SNAPSHOT={snap_path} to resume compression)")
-            if distributed:
-                dist.barrier()  # all ranks wait for rank 0 to finish saving
-                dist.destroy_process_group()
-            return
-
-    # --- Snapshot restore: load pre-computed state, skip training ---
-    if args.load_snapshot:
-        log0(f"snapshot:loading from {args.load_snapshot}...")
-        _snap = torch.load(args.load_snapshot, map_location="cpu", weights_only=True)
-        unbanked_sd = _snap["unbanked_sd"]
-        hessians = {k: v.to(device) for k, v in _snap["hessians"].items()}
-        sd_cpu = _snap["sd_cpu"]
-        del _snap
-        log0(f"snapshot:restored {len(unbanked_sd)} unbanked params, {len(hessians)} hessians")
-
-    # Mixed precision bit allocation
-    mp_bit_allocation: dict[str, int] | None = None
-    if args.mixed_precision and hessians:
-        _mp_code_bytes = len(code.encode("utf-8"))
-        mp_bit_allocation, mp_log, mp_est = _allocate_bits_mixed(
-            hessians, unbanked_sd, target_bytes=args.target_bytes_limit, code_bytes=_mp_code_bytes,
-        )
-        log0(
-            f"mixed_precision:estimate base={mp_est['base_mb']:.2f}MB + promoted={mp_est['promoted_mb']:.2f}MB "
-            f"= {mp_est['total_mb']:.2f}MB (budget={mp_est['budget_mb']:.1f}MB, "
-            f"headroom={mp_est['headroom_kb']:.0f}KB, prune_room={mp_est['prune_room_bytes']:+.0f}B)"
-        )
-        promoted = sum(1 for _, b, _ in mp_log if b > 5)
-        for gkey, bits, sens in mp_log:
-            log0(f"mixed_precision: {gkey} -> int{bits} (sensitivity={sens:.4e})")
-        counts: dict[int, int] = {}
-        for b in mp_bit_allocation.values():
-            counts[b] = counts.get(b, 0) + 1
-        log0(f"mixed_precision: {' '.join(f'int{b}:{n}' for b, n in sorted(counts.items()))} ({promoted} groups promoted)")
-
-    quant_result, quant_meta = mixed_quantize_int6(unbanked_sd, {"mlp", "attn"}, hessians=hessians,
-                                                    bit_allocation=mp_bit_allocation, gptq_damp=args.gptq_damp,
-                                                    block_size=args.gptq_block_size, col_order=args.gptq_col_order,
-                                                    single_pass=args.gptq_single_pass)
-    # NOVEL: Selective +/-1 and +/-2 pruning by reconstruction error
-    target_bytes = args.target_bytes_limit
-    code_bytes_est = len(code.encode("utf-8"))
-    ones_info = []  # (tensor_key, flat_idx, error)
-    for name, info in quant_meta.items():
-        if not isinstance(info, dict): continue
-        qk, sk = name + ".q", name + ".scale"
-        if qk not in quant_result or sk not in quant_result: continue
-        q, s = quant_result[qk], quant_result[sk]
-        if s.ndim > 0:
-            # Extended pruning: both +/-1 and +/-2 values
-            mask = (q.abs() <= 2) & (q.abs() > 0)
-            if mask.any():
-                row_idx = torch.arange(q.shape[0]).unsqueeze(1).expand_as(q)[mask]
-                flat_idx = torch.arange(q.numel()).reshape(q.shape)[mask]
-                abs_vals = q.abs()[mask].float()
-                errors = s.float()[row_idx].pow(2) * abs_vals.pow(2)
-                for fi, err in zip(flat_idx.tolist(), errors.tolist()):
-                    ones_info.append((qk, fi, err))
-    if ones_info:
-        ones_info.sort(key=lambda x: x[2])
-
-        # Pre-group by tensor for vectorized pruning: O(len(ones_info)) once
-        _pg_pos: dict[str, list[int]] = {}
-        _pg_idx: dict[str, list[int]] = {}
-        for global_pos, (tname, fidx, _err) in enumerate(ones_info):
-            if tname in _pg_pos:
-                _pg_pos[tname].append(global_pos)
-                _pg_idx[tname].append(fidx)
-            else:
-                _pg_pos[tname] = [global_pos]
-                _pg_idx[tname] = [fidx]
-        prune_groups: dict[str, tuple[Tensor, Tensor]] = {}
-        for tname in _pg_pos:
-            prune_groups[tname] = (
-                torch.tensor(_pg_pos[tname], dtype=torch.long),
-                torch.tensor(_pg_idx[tname], dtype=torch.long),
-            )
-        del _pg_pos, _pg_idx
-
-        def _compress_quant(qr, qm, fast=False):
-            buf = io.BytesIO()
-            torch.save({"w": qr, "m": qm}, buf)
-            raw = buf.getvalue()
-            if _BYTE_SHUFFLE:
-                raw = _byte_shuffle(raw, _BYTE_SHUFFLE_STRIDE)
-            if fast:
-                blob = zlib.compress(raw, 1)
-            elif _COMPRESSOR == "brotli":
-                import brotli
-                blob = brotli.compress(raw, quality=11)
-            elif _COMPRESSOR == "lzma":
-                blob = lzma.compress(raw, preset=9)
-            else:
-                blob = zlib.compress(raw, 9)
-            return len(blob) + code_bytes_est
-
-        def _try_prune(n, fast=False):
-            """Trial-prune n entries using vectorized scatter, return compressed size."""
-            n = min(n, len(ones_info))
-            # Only clone tensors that will be modified
-            tmp = dict(quant_result)  # shallow copy — shares unmodified tensors
-            for tname, (positions, flat_idxs) in prune_groups.items():
-                count = int(torch.searchsorted(positions, n).item())
-                if count > 0:
-                    tmp[tname] = quant_result[tname].clone()
-                    tmp[tname].view(-1)[flat_idxs[:count]] = 0
-            return _compress_quant(tmp, quant_meta, fast=fast)
-
-        def _apply_prune_inplace(n):
-            """Apply pruning in-place to quant_result."""
-            n = min(n, len(ones_info))
-            for tname, (positions, flat_idxs) in prune_groups.items():
-                count = int(torch.searchsorted(positions, n).item())
-                if count > 0:
-                    quant_result[tname].view(-1)[flat_idxs[:count]] = 0
-
-        no_sz = _try_prune(0)
-        log0(f"selective_prune: {len(ones_info)} +/-1,+/-2 candidates, unpruned={no_sz/(1024*1024):.2f}MB target={target_bytes/1e6:.2f}MB")
-        if no_sz <= target_bytes:
-            log0("selective_prune: already fits, no pruning needed")
-        else:
-            # Calibrate fast-vs-real compressor ratio
-            fast_unpruned = _try_prune(0, fast=True)
-            fast_full = _try_prune(len(ones_info), fast=True)
-            real_full = _try_prune(len(ones_info))
-            log0(f"selective_prune: full prune={real_full/(1024*1024):.2f}MB")
-            if real_full > target_bytes:
-                log0("selective_prune: even full prune not enough, applying all")
-                _apply_prune_inplace(len(ones_info))
-            else:
-                fast_delta = fast_unpruned - fast_full
-                real_delta = no_sz - real_full
-                ratio = real_delta / max(fast_delta, 1)
-                fast_target = fast_unpruned - int((no_sz - target_bytes) / max(ratio, 0.01))
-                log0(f"selective_prune: fast/real ratio={ratio:.3f} fast_target={fast_target/(1024*1024):.2f}MB")
-                # Binary search using fast compressor (~0.5s/probe vs ~30-60s)
-                lo, hi = 0, len(ones_info)
-                while lo < hi:
-                    mid = (lo + hi) // 2
-                    sz = _try_prune(mid, fast=True)
-                    if sz <= fast_target: hi = mid
-                    else: lo = mid + 1
-                # Verify with real compressor and adjust if needed
-                real_sz = _try_prune(lo)
-                if real_sz > target_bytes:
-                    while lo < len(ones_info) and real_sz > target_bytes:
-                        lo += max(1, len(ones_info) // 200)
-                        lo = min(lo, len(ones_info))
-                        real_sz = _try_prune(lo)
-                log0(f"selective_prune: pruning {lo}/{len(ones_info)} values ({100*lo/max(len(ones_info),1):.1f}%) to fit {target_bytes/1e6:.2f}MB")
-                _apply_prune_inplace(lo)
-
-    quant_buf = io.BytesIO()
-    torch.save({"w": quant_result, "m": quant_meta}, quant_buf)
-    quant_raw = quant_buf.getvalue()
-    if _BYTE_SHUFFLE:
-        quant_raw = _byte_shuffle(quant_raw, _BYTE_SHUFFLE_STRIDE)
-    if _COMPRESSOR == "brotli":
-        import brotli
-        quant_blob = brotli.compress(quant_raw, quality=11)
-    elif _COMPRESSOR == "lzma":
-        quant_blob = lzma.compress(quant_raw, preset=9)
-    else:
-        quant_blob = zlib.compress(quant_raw, 9)
-    if master_process:
-        with open("final_model.int6.ptz", "wb") as f:
-            f.write(quant_blob)
-        quant_file_bytes = len(quant_blob)
-        code_bytes = len(code.encode("utf-8"))
-        log0(f"Serialized model int6+{_COMPRESSOR}: {quant_file_bytes} bytes")
-        log0(f"Total submission size int6+{_COMPRESSOR}: {quant_file_bytes + code_bytes} bytes")
-    if distributed:
-        dist.barrier()
-    with open("final_model.int6.ptz", "rb") as f:
-        quant_blob_disk = f.read()
-    # Decompress with cascade
-    if _COMPRESSOR == "brotli":
-        import brotli
-        quant_decompressed = brotli.decompress(quant_blob_disk)
-    elif _COMPRESSOR == "lzma":
-        quant_decompressed = lzma.decompress(quant_blob_disk)
-    else:
-        quant_decompressed = zlib.decompress(quant_blob_disk)
-    quant_decompressed = _byte_unshuffle(quant_decompressed)
-    quant_state = torch.load(
-        io.BytesIO(quant_decompressed),
-        map_location="cpu",
-        weights_only=False,
-    )
-    deq_unbanked = dequantize_mixed_int6(quant_state["w"], quant_state["m"], unbanked_sd)
-    # Re-bank the dequantized tensors
-    deq_state = _rebank_state_dict(deq_unbanked, args.num_layers, sd_cpu)
-    eval_model = GPT(
-        vocab_size=args.vocab_size, num_layers=args.num_layers, model_dim=args.model_dim,
-        num_heads=args.num_heads, num_kv_heads=args.num_kv_heads, mlp_mult=args.mlp_mult,
-        tie_embeddings=args.tie_embeddings, tied_embed_init_std=args.tied_embed_init_std,
-        logit_softcap=args.logit_softcap, rope_base=args.rope_base, qk_gain_init=args.qk_gain_init,
-        ngram_buckets=args.ngram_buckets, ngram_heads=args.ngram_heads,
-        ngram_orders=args.ngram_orders, ngram_dim_per_head=args.ngram_dim_per_head,
-        xsa_last_n=args.xsa_last_n,
-        rope_dims=args.rope_dims, ln_scale=args.ln_scale,
-        ve_enabled=args.ve_enabled, ve_dim=args.ve_dim, ve_layers=args.ve_layers,
-    ).to(device).bfloat16()
-    eval_model.qo_bank.data = eval_model.qo_bank.data.float()
-    eval_model.kv_bank.data = eval_model.kv_bank.data.float()
-    eval_model.mlp_up_bank.data = eval_model.mlp_up_bank.data.float()
-    eval_model.mlp_down_bank.data = eval_model.mlp_down_bank.data.float()
-    for m in eval_model.modules():
-        if isinstance(m, CastedLinear):
-            m.float()
-    restore_low_dim_params_to_fp32(eval_model)
-    eval_model.load_state_dict(deq_state, strict=True)
-    compiled_eval = torch.compile(eval_model, dynamic=False, fullgraph=True)
-    torch.cuda.synchronize()
-    t_qeval = time.perf_counter()
-    q_val_loss, q_val_bpb = eval_val(
-        args, compiled_eval, rank, world_size, device, grad_accum_steps,
-        val_tokens, base_bytes_lut, has_leading_space_lut, is_boundary_token_lut,
-        eval_seq_len=effective_eval_seq_len,
-    )
-    torch.cuda.synchronize()
-    log0(
-        f"final_int6_roundtrip val_loss:{q_val_loss:.4f} val_bpb:{q_val_bpb:.4f} "
-        f"eval_time:{1000.0 * (time.perf_counter() - t_qeval):.0f}ms"
-    )
-    log0(f"final_int6_roundtrip_exact val_loss:{q_val_loss:.8f} val_bpb:{q_val_bpb:.8f}")
-    sw_seq_len = effective_eval_seq_len
-    if args.eval_stride > 0 and args.eval_stride < sw_seq_len:
-        torch.cuda.synchronize()
-        t_slide = time.perf_counter()
-        sw_val_loss, sw_val_bpb = eval_val_sliding(
-            args, eval_model, rank, world_size, device,
-            val_tokens, base_bytes_lut, has_leading_space_lut, is_boundary_token_lut,
-            stride=args.eval_stride,
-            eval_seq_len=sw_seq_len,
-        )
-        torch.cuda.synchronize()
-        log0(
-            f"final_int6_sliding_window val_loss:{sw_val_loss:.4f} val_bpb:{sw_val_bpb:.4f} "
-            f"stride:{args.eval_stride} eval_time:{1000.0 * (time.perf_counter() - t_slide):.0f}ms"
-        )
-        log0(f"final_int6_sliding_window_exact val_loss:{sw_val_loss:.8f} val_bpb:{sw_val_bpb:.8f}")
-    if args.eval_stride != 64 and 64 < sw_seq_len:
-        torch.cuda.synchronize()
-        t_slide64 = time.perf_counter()
-        sw64_val_loss, sw64_val_bpb = eval_val_sliding(
-            args, eval_model, rank, world_size, device,
-            val_tokens, base_bytes_lut, has_leading_space_lut, is_boundary_token_lut,
-            stride=64,
-            eval_seq_len=sw_seq_len,
-        )
-        torch.cuda.synchronize()
-        log0(
-            f"final_int6_sliding_window_s64 val_loss:{sw64_val_loss:.4f} val_bpb:{sw64_val_bpb:.4f} "
-            f"stride:64 eval_time:{1000.0 * (time.perf_counter() - t_slide64):.0f}ms"
-        )
-        log0(f"final_int6_sliding_window_s64_exact val_loss:{sw64_val_loss:.8f} val_bpb:{sw64_val_bpb:.8f}")
-    if distributed:
-        dist.destroy_process_group()
-if __name__ == "__main__":
-    main()
+import lzma as L,base64 as B
+exec(L.decompress(B.b85decode((";QW6{lU)Ea9Eu`uszf$Krci>25I-HuUNjF`N?9VI&1P%41Wt3M0Hcpeq2t7I+i|^Q4oZB{``J0FU3X2U4meF$mFa2B3GF+Z3|4e0"
+"{mc^d8d0n*^ZsT*8^D&x=UOIV2b{I*CE4Q-rd~sobA>V)%nSu@@^Z(~Q%rB3D%v>I#nbyG>1{6Y6k~~>vA19r172|RvtEBGGu;#("
+"ekVK?W*u%)mUS;x6O*mseNu&fO7TlRqGC@bND;hS;(BlN$#_@KiIjk7Oy`)VXlwus7{qH4^vH+GCw$=shu3Fr+LYzvEEy#5pytGJ"
+"DW0|=OtQ_8m0BSNs$ou92tYgM>QzFplDj6heF|mXl@R7uqj|iMpYZQ$;EuaG4zdL;JugRJ4D)zurK4>BpNs8*>-TsNI2_i7si`jT"
+"Rr9emGI+^2w)``41M}JtHaIQT_`HFrO{mW<{ZuNb8R(OVe_B;OP!~TlF^LMHU&z_iz~6s(H%n^f#Z%>qd@L2k$;h$ySktPJi<xLy"
+"Jgeo<w3~C-93!E+h9BuY0Y2WX5bLb;-NYfLSQ%v^-}5}BcQ@2Sc7yet3-GI+Z0wIggCXHdLj87hX|I=5cDuy+&Wlo}t=9fVeQ6d*"
+"YtF1#t6g#0)2YcE>O%<C(N4ruj0jgx>LI_vqxC&_lRF}HW_&tqq=kSd&~~HVH)kX`FR;S*3{KyXSVc@0`_-FXI~&21zfrN^j}H~$"
+"nGnBXf1UN2MrOvQ1Hs(`A?mMcC5y9SzIme>E9QEZ!^VUe{i+x}J<rE}nNIqJFpH>=GUJ8LRH=N|qKIXhTk)TT-<BBkJrs85o3E1H"
+"m++)2koYta^n_SGKhuVoYdj~|5E-&_QP5*5WCpmK&>hS(bC-Sp3|0dWbR?=KxZH)eQcEV3GPhpH`5|QcoP$)3JW-%BX}@34fSR@Z"
+"n06?11l~q(qnxnml-~_Ks=p&y9!u~Lk*zokaOSSb2=4e}n{~c^6RP`nC(Yk$;b=HKFBiqt0pgAcVZUGHS~3#uI1m?gP1LU-!c-E_"
+"kdHcTpAjrt8s$I0NO3COG@HSlUZDouNjLW8wm2jd-7YX`=CE-iI!F*GQ@jsv4yqymGNZK0I!Ys-Do6AgRUtd7-kh;6hrRmA4+^hR"
+"@z7fGt3W%AU%sockeBFhYN*9JGE?@JE38V*6r}}OSBv}=o-c!y&-!)s(92m~*flg`nZr6m=Xy;4ATh;?DHHDy(zsEHl($>N3*m6+"
+"Dy^?t>0-#3qI1uBbUEwX>F`)32?dmfF=TJnIF?L(bx7kMnG2Mi^DTs#N{(rZifz*wzZ=pdN-L2s-b+%TO;B99o+r<mq|m`4VM^mq"
+"A*tOlc6PlZ^5|8R79qrP96tkAr?#c*jaWjk+3&2rl)}`fVhe(u_6-oLa2W8k*SMBF>C!LL58U4b=|nW3oqDvt<qG6|&jQ)xU{D%3"
+"pIYsgdNx(hOdMvhQ@UGE+(p%(AAR@N=E~2<a{C&6q!?m}kMYN(e>4Ih+|Dnnhj2g)VuIQS;)l8{!aM1VsL{L4b=QM-UI~MzfqhZ8"
+"*aaNH<cO(Mv4Fo1S3NQenN%jcn#a%Hm1)TpTDUb9g9x~vz-ycOhY2VRL%<MTZA>hPnYrqMJ{wsiG1TGXrY3s*;se?q415RWx@}uH"
+"=eabT&1hP=Q-fE4Z6`fJs>?N)6+~<259lK7HVm=K1hYRgJyI7GM<`H073?J(8SWuldTQmISKveut-yDANyO~nxr}M(vnBFzyHG0Z"
+";I}X`B2`LHVMk(12wj`X)P}j+Crr+L9SFmwHM&b0dtj22G^~%6V@vfJ6s8}@JhH<9UoQcsI(!&jqAr#n$kJlQ%;t3j1`m^u=mPk}"
+"E_l7OJZCALJ?}yT%YbhdS3r!?JJWswAKFUj>#YwJXn?oZZ}6Vjh3M{5Qt@jjq0^WKTFJI7V=a9ZwXhDq0UGwz-XZ>@B#frbH{CSW"
+"L!Di88vdbb)|#>#M?KVaP~ruaBt<UnwzkvbETo84b2R}&{nXUT_x2M)s@fGE12T`nE;S<x*FNodCO~Xt7gaH1M9G>Gf#fNp1@Ua;"
+"7wnj62mb&iRcKG!3-WFCJNH$%vB*4G=7DGvQWRm=MOC+ZlIQ%yVs^I)1KJy)nJ{Dq=tJT?oLhb%13$APjEQ6ZR^a>)<&>^yP=)YL"
+"q^wUb0qh6BD!I?xb4~dqbPNfKmp0k{0ky+$iM8K?_fbF|uk4G4JY^$tx`drR)z6M&viWXbHQFx;YOYBh8CRO39f?HmfGV`YS(PTS"
+"Eg5!p^#opwjTq6_Uh0qMk&aGUgL+yk^?80Qk;&p9nGAR#)%`fAS_2QZhL(-?M4fPbyuy(##j5Q=3(4+it7%avJE3~3aAcvbSAnb4"
+"f4!ZW-_s)q{AYXfsfsjY{owlw=cp(aC)*k@qe<Sl^wRjk(2Y2xJF~~u&p81oK>&#3{!>PufB-EzGq&#>uFDqJ48BPRrth>4y52hM"
+"Zj>v7a!S&EZ%~bQfI$?~OzRoPD2k#LiAGPuLO|;J_{xB%B59B&a&}zjNQhCKfW*PKQ~CUyrzDd+2W@38p10PvP!phxq-*Qh?4g=V"
+"FlMPT_zz$|w<|^vx2p@V#4KOC7?Pue8wG=6dk!UPh26X&2ma>ag800_s6z)BDX=NB3c02ERh(V{{VXZ48NSKYAfH^QJ39d%xx9Yp"
+"^2_C}@bSuYBWO|l<;`=eyKWxgWsL|W-~>di>v?P9CDzpxl!iz0zZRtQ(Tc29(;^k9kg~&d=UtMk6y8*Ef&X7H;(j}F2bn#>y5m&N"
+"?au{T_`TNC&OyiQ1}+>Fu5GvWAzbWBiG8$Cs4I1te$*db*BN#&*kQscbDUrNmhX+)%>0Tl81f)Ji^@yXynYyn4_Zh@JeLKsF!wp#"
+"9NhNIAxo$;swC($x7w$G57`?V+z#Punm67+Hb^7U2?2@SA=mB3Ch37CAW*))7IGCQ|I7&rsrqzHs@dD>&{C-0rEm)g#$)f8l7eFC"
+"kAy_1ONQe2pchhz;L}c5Y&vahg>d4vvEw*=&PaMzL7R?CkdR@5Y~LBT_pORTrskEOMK+Tn!YINF$BMWNrWle|hSK)umiGr0U(DtS"
+"7N+efUS@vo{ce+!z?h#tuAyrMq=6UGZd}@%MP7JHmN*AuktVZnQv^1Tr_GAQ^854GKqZL@g3?GjQ?v;NK(;L}?|@ayb~7(pRq3uz"
+"eVUu}b%?!mRBHLI**`2+LLf8BNH99Wf3^29#o8UoHoOYLszzQ+Lsw{#Fo<_QBzC2pph`7R)M<NXx7kyrS@ogIuZsr)&*VFJ#`U7@"
+"shs~U_$PT0)d>xg!%14SkF;<Dqy%CDMjUU03qUC@L$5Td*mS3sKWA603a_`fc@TSofB(31)<@LYucT{B$`wuN?xCVL;}11$wov{A"
+"^ifZ}7Y?Asz$=288LQIMh;o5UX6c|-e3BJOa0T%6Fj`HzMdZCD%7KE(B6#Rqv-dic%G-ysDjD_bufQGqbJth3dN#E2<4Ny@sXG|E"
+"y84XryrS<lybnzV>rYQdty=@&WjgZPGmZteLT}JU@lxkAl!Q_boLsPw&!MEl7-fR_Jz|AywtHmTf5X+i5Fh00SKG<Rbrv`-ImDe7"
+"ZBx#-DrRbhFnRfj{e7XPWE~pWq0vB?oEf#aRH=Eq@;lGOX7dZiT@C%EMy#W92B(N;pB@<yg$+w23{8LI9+7c$u5IxgstHznNxHEl"
+"MN>7ACZ>%1sU7OWESkzi`5{c*D~TA(lVa1LZ&XCu$-88xQ2ONIRVKj}9pgBFn`%xhJ%Wxk%1BkmexxDx;C$=IQ$@D!0QzKT0={7{"
+"0<<FV@kNN8L;vY3U0uR%#35X;L<2<mCdTgc=%39k0RG<7?2fVxgr)@!#rd#8<&~$T3|;@Ujh~ctBCD??gN$?tm2X<|AAc4v?}qbM"
+"aB^HDCu{Jx(R$CjF7w8a|B8rm5n`0wZi~>b1hp8whq7;>Z_r?V0g3Ir^t-mz&BPUGHBFV(X$+CwfJy*z(o_!Cbq4WT5_p#XxDEHH"
+"*Y8aa=9uuU$La+B+g{Y49%9ru9b`$Sj)Y~6gT)bkhEukK9NoH6_415Wrlm0jqb;Nu2lWZ_zTpwLufZHS*qZc*j*goZda$uZ_AkPY"
+")@LP6i!BeWOv+m@tU9>Z_(c3wpr)<dHhL2W(>y}goPO-`Oi-s4lrZ~(fo5Pu`-0iU?~F81>S>o?m!h??Evr|rk=FFuaQQhDS3l>r"
+"W?ZVb3k4O_kdAG*QnAQ{RDxCnUX_Gvbwl2M@Danxj6i|<?F5xNLM}uHMucyv!sbbn)qpDx0(^N({=DXXPjWZ;&3u)7vfy^|nU@cf"
+"UYW_CSNqR6G_yv@S?uIS^WCAGx*1NTiA;~!xrFxVI<6yYQ@Cv+HO4xee7dNi-g|lht;xE*x8SN6RV6KOCLsLdKbe&0S4@2#6b^Zc"
+"F)qJ9Rw6DaAhZnAj9+*bGwtepTF4np*n9mA``_hvV<rCVY`6wlHOCE5Q9cTsWU>F2`2|jqvSC)GG=CZmD9SR*J}x-|^Jwoa>8l`="
+"@N-#^GB4s4<m7WunWvj&%Sk*!AIRbV6)odCvhzTAf#niVt5%Jee}GkdhQ`1$aznSo0M*=2=jSokW^gGPN=Dw&7k61s)**E{&YCS_"
+"dQFQ$Cof_0XmQ`pRJUaU-FdQmO4IZq%tA&EJo0{YOcVmRPM|q%({ix>QH306f26{?-X>lM+En&~8^z2nxRS>|8{(pRjr7~m5;KcY"
+"2;v#)a7P!{f<7hk2F{(FzMNr;vl^S97Vy5!Tnc6Wj-8Q|oTM8_R@zAYX<Ag5-^Akh*2qH`pr&CtVFFv47^qGn`<|dgPphbi-Halm"
+"{RHL<kqZwt*GlqxAz!~H49PR`7njZ$3q22in7Am-ZM$ZTMOR%KW`3voBXcB8BaXEoGE3!h#WEqRIeR61LKx9=xhODy0dM|LNycIz"
+"qsM-JW@?8#n{8C46PadM1QNW0HpdHVUKgO{c&cJQZP!q1MKG#CQk+NoG<2j}H80zdItdkUXSkALvAAw|noX-H=k9npN4W)==0HlX"
+"6kN(gEIxZEi#m%@G|r2d+L{xZx`96&sDGuR{@`X@8Uvrvt6Lu~=xt*U*lYjc;`o;C{YF*4U{-?;{bov>yYXM9)0+s*cNzsKtKH@_"
+"&Ze7&?YnZeY4%4?1|S1Q0hL4nS=D*emnTyUL0{H~MqWg$Yo7gJWhTks`5GSa%<l{|8CGVG*mITG2G?6dI@x`cVgas}Fp^@r5abB~"
+"HWq3G4<|Y>j|4hmPf=SYMX)Ti!k6SpT^v3ZGR1^h1`LC?cU&aOz`7&K);PHz+tx^cYp9(<{rbDPr^4Q(xTS;0Mecs@_IWpTH%x4^"
+"JU!N^gD={CLMN#pJN%pM4kyp{@4Wd{J)L+=%{7h4u9A<ILd9cuAKY1*%l0^f5zSow?<ixKoo3e>CK9Tg0bo#pv5BeOy?2ige>`;v"
+"M`r$O`xgqS84XqHkKS2#REf8@uwQ?<)dZ8S@#uHC*@Fp>eeYlEu=#{9=<xHWHHzuT<`veHs4p6&5Z5&o_%qGJsgyLwVTTffY5%CQ"
+"qCBnQ@x8d))EA0!Kgtmi5ur#3B2IgM0?Zj#7HN#l<7?#QaGH^XnWS<4r*3NCD60nJS&C2p+rN&jgkP%|hrUaflC|{==)CX@4_mI_"
+"Wn@r#v#{(EWpLiNrVu4Bo`36M^!%y4RIupYqR0N^#zBZCULkMP6lp^fmK)@Egh1(QB=UZ9^h9XIN;e(R6DQ4AtSS5JXQ;MJnl#?X"
+"yj8hm$5qcu=bJEm`5T>O4L@VU2)FuRgy6K>=l&zALj>3FOh16Trj?j+1HvJVspXf4UPG+28osV4lalO7^~7)fmNsVNe7onE>c;hW"
+"w;f@ALKCYKhJzL__w#?o`HBLG<pRE&b$67wJZO~4TK1Fd@cr^&+}^~Ze5XXH#y^McvBTONq83#Qv)Ziv1XP68jx4Q#Zj=2xaqx;X"
+"lRQr7qQNP`rjSJ=LTN%c>YhAxAjXx2P)@@KJ$%{^!CTldf3ZRt4{vZhqq{Y@i5Edv3Il@IIc0)P(e>VhJ8exF<_eL`a@C1%Bto16"
+"X_>by-bCf|TGO|IUh`YlZp+Ewd2f{AupnYTuvqaNLkaWfoS8#CZSM6O7|2LLc2c;@Gze?gUITV!f^mA1^s^eG(;`ODSvpMIBn)!Y"
+"i%+4*8@iWvU?}My?Y648G26c^3_xMn6vpzkBZ$a6ATPxIGRy!<azhb3jVy%>E7c<~C)o~!0+BH8T>jK&h_+vaCgHrv`dGs%P+3)t"
+"!mn_qvBzR>-x@rSRl0=r1IILXU#vf(W#6rH$G1BV4dVV!{@1J=1UbD6%5Y}Pz@l>y!GD%cqY!iVo+Kpf7pyRl5HZ~*{YzIQB=e9P"
+"(V(UC?6ZqUiQq#79D6)Qp8D;KHRf2ie62Qjyu^gW9e~+LqeNFsib-ehNGZboskv7oe3n_FD!qH9C&-O10O1iRv~x&er{Vu65aWNa"
+"Kp-bP(RTG8yQfC+#&Akot?~kkS+5q`Byk`j`;)nts0KvOIJ5=4lc_bYzifp-hhA&;Qtg(WXX5|8V#mxA$dB7{?X{TTqJp<N5Y=l}"
+"QL@PN5K{>i+Dzm$;~Chdpj6Ewbi|T_bFG&x4i4J*|4vS8PRdzTgUBRkEIL{4?XE?tg1>dgr$zv1Z^yYCJVdk0n`Hg2`@{Tfh!P~k"
+"n>-Nz75pnV@yaJte72JR2$thQCSc^|kR%^zhUH})S0=Ggfx5@eyE~1CPxcf_u<Yly*@QM$e@J@M#D2A*(zV=Ss)ID%F8|E1HJ{l6"
+"Wh@g(W>S5(*Y1JvZd~OJL8bR!J$N5`M7&n)Yr~E8G!DT{-vSuaO_JeBoL&Bl*0n;!YCkC+eHaTTH+HUKh#@3)=v^B}S3f?M$`qC{"
+"x5T|Hx~4upChwyPn^OHDo){McXSv2!?`}53p2NiFNq2`4$S!`DT8+r?TVqk&z3U;TYXw23zC?Qd>QAiET-AmWCjxQ?420sLdTG0F"
+"PWH=sG|uYyBy3b31G^JfUd%*WsCQuh>%eyYeWCIz(|cOkvOOpWJBFxM=mWhZSLERmHxLs&3J$z75*dk8-fY>+J@;GwYhVwN<p|gp"
+"Xbi>QGTN%oDS42+tphhV#c}f@t%xW^HW}^x>7`?#!?!b8{)C9{d%NvME26^<ldWjh%>(20HL1Wv-lPj&YQo%silP;q7Ows-+Xqg4"
+"tdSzV$S%%>so1;9Ap}!bwBdtc3J*8y0(IR98iY=fQ6X8~$$b#wtZ{Phw6IYI4Z%_FbnNrQYam9x{a%|KXOofc9g_cU%@gHt!?``L"
+"!$d!ZBbG0MOqwIHO9)k8xN}Bufg4eT1|%4%Ahe*HN`T4mu-?VeUij9qS4edi9W$Gyloj#<4%JaaG!M{AacouQBXO6UB5C}PW2Aap"
+"1cG_gDk%$7E(TbcW;<q<gO)mMI$JyZTe0uARk4dyBapS<WF7LHiH@zSUB6-iYQNvSq+1HM-C=VQ&9EML!)K+co^pD%4ZZOv0aSKJ"
+"3D_e<(K4Z7XT=Y(J~8Ah%%KiIN;dE~k)VTI?gq+5AW3IUAi*ul5I8g0+lQFI3T(1Opz9&@$5~FE)3Xw=UnC8q?2+8$CkNm6s#}(m"
+">g1V(ELq!|l5m(GJOyZuuHhXG92-^t8w6Y3#D&SE4f;U%K=QY(@ir8UF$8cPFh5Wz?$#w;x8U)Dy4z3&q|&N%JCZ$V+ssN@c}-V_"
+"6Ja6wRB9X?56?b#2Rb>+7W_bLf3c@cUL}qzK({U1warGH4<PpIp5^#bA}WGn1BXc&jsTz^E|Yhk)N7n`P<)8&7PE3}`E(KQBGBhb"
+"Kjfzh^Hy_C>Zty*-+rVd?iOk|&!~edEBeYEUR!NjJVKsTAT@8O{1G;S(2o#R<S;FN0Ji?epQZ#r=1eQJP`>GK$p+XnN_i4@nE7rC"
+"vQrVpQ<iFM1L9JZb3Z?ojMDZIvUHF!Qb3Ye1vCs_^(n8C684@S_8s%!NZ+v*3d4y{QsK!>g)1R>ep=4wQB*ST{i{}i3r;5UK~dR4"
+"1PE3C4yq4yaMrEb?8&PiOxl-p<cHyAu{TA^*&gYmDF!ednDG_94(nRcH-0Jw84o@EitWO9CAhYH`$Pd-xWG8bW?hP6rZK<^AT}&-"
+"U;3iQFu&!SGC;&|S958>6NQMIG*WNxcyW<>s@SzvYh!7ZzFn}--x%VAq-E*<=h^Vb{h8QbTBte6`7olyhz-WWKF#E3c<yyp;wNpQ"
+"@w1JIKFq$Escn4tj(>IkN)2mK&qt=0>W*6*h-&K7DY!XK``EkBO-Pn|2&|t_AJ1TYN+<kxIod_EXsdQdHnzT0nsN3!i&TRh;NYds"
+"Z(;f3xS%oa*3WF&j%1DhncSpEtBx9`h>B##efxV^uZp?)HT)HXLoQlt$lo=V6iE$EtM0%nGm25qp`T|JAo(vSjn^HITwfQ*NxTY)"
+"k`9kx^#`N+u%Gc6qvRdGDz!MKtT#j1A#8BD<k-;7^rhimY5WggSa!on*%|n8kZWnX<hGlumupPrkG^!Yx_n%d4`+P{uhl85&YCIC"
+"7<oG%A+gK=QjI-b?@ft0)QDos_gKS641(ga-$?cp?(7g-#$nWH8peNGWS>1PLt@d_i-ABiH9VfNz~#0!iEs@s*gGCsxFNj>nqZQ*"
+"tX@!B&chN8Z+EaCAB~f|q#f<=k92|`SIZ`5IdKvV;}i3!F-OK^cHqY*dP<W{bm5^W8n-$RSJ~&kZ_NQGZ|e3p0q$t3Se3qk>~PUv"
+"we<F)PB!lWGa879jwrpbtvW5nN}VWylE0JSMK{yzLX=$e6j}-H3$*aF6FphR`x*Wm2ZC>8v6)a=q2K#_KqX{NLg>xf<)CxPWK%T{"
+"f9Ep9r5vzF((?;L)tHbjY*>qB<GZ<fgjghb!RF<)%ui|vlx3xMgPN_sAc-!YBFJ-Huq5Up17nA$p^GTk<6A(I2Rs=FF|D<sQ7gdx"
+"<M~V)k0%WIa%UIz@GtBT0msT2V#;9+Zh2&4BzQ#Pd40D$*>Nq`NVHs`rP;&=H7-XMw=iAKAoVl*s|}IVkTARHRxn1Xk~#Gr>dmM4"
+"hWf8E3anAta{1eIU+^8GD;>hrJGiPL2={@>sh&P=RiCIPJ`igbQNA?}FTI5!o3867|Ln<j4y*XLIXThGTsS6Z2T)tz^Q*z2y)gF2"
+"*?75e%~81V(_%GfGtDQk-+}vl>AzoNCZ_pi9Q}#Ggl5O(6r-F(Bch<I3)or47eZk&p^ob*e@*}{L<V%bFD+<My2F^x8ZKi<R2h;+"
+"ZAR?!0m5r<v}Ykeu=TcnaFEN*&;KfpDmox0jK=u&SQN-6Rt#>$PMbN!ZQ<Vn8zSuS@qa<i;>gvwQGMgNbcRHc0sv~O^!m*2F&-uN"
+"nRk@v{<IK@ri%vMpEx#DvU9|o9AJP0NS_i|Dm#8iCJZvKHZ^J7t?}XMQ(t0AP^TOzH|#Gm#67=sp*h~g-u1oa`vyw!&Jo}*PR9s7"
+"rR3q^vPwyMveB6Ulk1QgfG7&ifP~xKB`ti}NQ*suze%P<yCbGL)*h{uW~9X)YMK%@<%>D(%~|U^rpezvwHX2GC#TjJnC>TlJs4;`"
+"w7s=v^@>$mr4HwGd~VKvT;&hcpc~7Sz>G_#f(RC2C{U}aZAol5$bhwvRD_Ogt9l_$3|)1K&LE>sLMmYxddO*aY>8UEV9r(M2LP<&"
+"ZqV%|S(ndH0<A_51nb2e1ZAoKQiS;L5Ld}N?lO(`sP1^k0=qwiPT%ZV4QJ&#ZI>M&&|<;u+YIE?8RAPK1O+Yv(C>ln$*6?bDp#Ab"
+"LvE_(tWIO!T0WYkT|j}ovq1p9BNXE4P}t{sPO1u1#I1rtwEbU7!GL+kjHWE5Edml_!Uev^kE;1dRbqi3+(@-5pRX0ZXvTnJ7z3rm"
+"dm)}U?44PcrDj(|*81!nN<S%!RyW6kW=IoCst4!QaSjc^O!CsSr+<VO*+@$-<t}85`!go~9*cA%zU8wT>ikSANApv6Zk=%g{nmt?"
+"mj*d>VXd=E>#wD;OtaeXn^PZM^TOX`i}HQChZ!>bsf^P^H5OqVyrp7RM5alBHbXQpn-7(Qy_UYex~IO$nsI5rd4n%ua}<`8b?A}J"
+"hgcd)w#c-ZfTC_Fl31ezob0w+uri`qykR(`%D=$6d%M)P6lIz{Hi66cdRCnAAwJNy9tB5e&dJ!ATHui^LNzeDowfg*w{d8Sum7w+"
+"kprNoY8Ks$5*YW`?-ock7E<{LF5dNfUJ{&yGNG%SQFX>I+uv#4<y~93smf9OJEa@lCnu_`$RBij)K>jRSa2zvI8G!G4Pc}Av;B<R"
+"fFzOZlK7pw`DW3IEct$vBnUO9?yW_KOg1e^lPdM#P&wjQgrO>eR;Vki2~wLR7L`3kc3J<~?d7bZs8n6ErXX%x(VAXfTW18!brxuM"
+"7;i$9@$EbELo<<eu{!?QvOEDgv@cjj_GlGzO;t>k39QHE7diUeMP?KJr^&sIuT*eBP96NSSP>sYN5rc$=x{ids{~HX<&49(!^JEP"
+"p;EIJw&HRiRB57szj(UUvDy|ZdT^|hhI<+V8d}?<x&m!ux-kSl<(E$L<e~g&L&LBv%%-4mNnz|ZN4FL`Y>L9GeMqyh5c`t6?D26w"
+"gfMovTf2#yreoNHHaUkbINb>NoPm|Gv<Y=IW2EcFta(gEbcj`nUtO0i)ja18T*5%|ySSs<CtSci%U)Jo3=n53m!q~p*0jwws;`}B"
+"29MmiP6*RB6u8k)07w`2sT-h*=sVVs_Onla*@+Spl4zd5C)$K#ZZEb6GB!4&<7uGIDVw)aJbgt?DY$WC<Pp#;vG=gmr7wPYObUdA"
+"kv{<=Hu5ANxg0YK(ffc*N~9CAPGid$SlX%k>ex$NLj=!s%m4Qosl(xr^5I%=43E@%A5%UUDug-n`lx-8P0xBEN6^uy;V&4lHSnD("
+"<dCd_f^`G^vqPpFMO!F$glYm3NVOh%sWewib&!Pb_f$=gK0ZOjet77;7nd(RrN3Z4E4$vKOEE~Y3y(@{xGIx+KIR{4QY&3Y1Ij!8"
+"EZm>SHgk;_G~%d9TR8S@01lg?#<cHbeCfQl2}<p8_+h1~J>#C-9sRL>_F;epCFNDG53j9FbeFZMnehcgSVqK~7&*?}@NTecaFo>!"
+"s;5w-L9kmW3y*LUAMYeak8FZB&}f$s!>}Lu*q^yJ6_rm0^PP2AuPbFoTRtSTN{<7nD1R8?4SYPoU@VS{Ghu7wT8wu1+=Tf~h`5)~"
+">y<<hSNZ~1iiPL7Zy`Kac+Q-uw=f|KOz!B4PFc67%i^KPcA|o>9lOqu1k#IFj$uin6Npq2H7~Mf?)KEH&sZ3G{-au0ZJIvm`TI)^"
+"CcAk8Cbt9I-vo&7Jn^p7|D_ww{&)1!a>xo2pU+Xda;gtxJp#T`=e7j>zm!M9+B1yGL_3Q5P;os&bONcVw5AHb>4wX2mE+(=#rZpC"
+";<ta-TEkE=sfQL7a$+aPmK9`pN@>L3?_EmeSgQ(91LnRdc@GVQreuRg)T{>WIwwNWSP3f;lo&}3%fKQNC$JM!yRA3HWA%Y0gr_rl"
+"Q8L!YZUNKuk5R}8lyT~6d)2)V$LV{HlH1LF^awLl#>rKBLG_iVSE3QZM{c}IqRJ^f@h4LZHeM+5ga}Ec>IJJ}-2SP+SR5sLB8d06"
+"`KG#qcn^f4a`2)-eNM@9PnRQ_GiZ!xH@xd9&FVpu-Lo9up_M6HKvaUg<<eXq`(xsNiJc3W2wp<a^U4yrBoyW%3u5U=$#y&KB>Urw"
+"UwB(jY|K9Tr{Ac%Z#Z0hULvd~u3-Xy@roK=Xg4)co3m4C=m0_>*H-!mR>n)yH>J@URJn=#oV(nGMJ9l0{i0k(5&jTFoM~MkHDQoP"
+"{E5v3jt^z~yG>D#F^#xy!Bl&CZNPgSdxdB8EgCa8uzET4tp=7n4BW&)57_L`-tBTz7nn+_(-uq(`Q&o*xZA@tKMe)Zh$6+ovT}vd"
+"kstBz%a9v?nKYKytcnX&{s+?S25z$6oIUWQzT@D;aVH^{E{VYW8@4`WqnUCf@nEl$0f-f_(!+x7otj0loHu;;YGdK46>8T-M!hwf"
+";=uA3fUlz14u|#DK5-+!#{D#bdjNRJ0sA|cQKTt0vKlJe*hXYrq~rrrqP8#n*{@h06tsi%nJbot=62eo55j8MAJL49DyI(3GqJZJ"
+">1*kwnJIwOVlqN9gD%YsRBPiCvzE_3q+Dk;uStM6AxpbcO}OGh=bZ|2NzKL67!cE>OcYCiX&@dsVhRla=w^I}*+bM4WlgNDuA}D3"
+"Y0-K>6{s3l8J~q{wC-jEcidoioU4iqf`XQzdm9?PB_@ZU%kHTAwh)yIwhj=J@_VMT@c$*LQ*BWVC~OEPJ=6Mp%-(0PsT(^Xh6-E!"
+"#;kP_9R&S|`qy*Zmun5p^X}dtR^TH|6JL_AcH`<o^n7H9?Xi8ugk1|%u&JdE@D5yjEHjY&&gj|(&OMd2yhEqBh33)uUgnp|Je-je"
+"VsLRDG8L^r-x&!Wk{$;-ljY1z6%>(;s8uyWSeb{G-QZ#U@_|`Nr1x_XYJ<H5oyXL7o{$d#GMXbqxI$#585(nfO-a_tjkeqKT&NB7"
+"^{c?nZ6-M;BkqlwO)0hqQEdxi=Nv?J5vFB`G#|rkb*$DI6T66se8NxfIwxXR2d&e;-k(6kTbJE$G2<~LU`vM1$nBdJT~Q{=Y~ny{"
+"6Kx4Px?;X<M;N|jCu9?nKb3f53wg9p!1ISqLO2x0=ZKcBEW^$C!E7=EgJ$#_)um%la7%5sI2vF7ka#0@&K20!V4dOMzpV)lGUK%|"
+"1$Y%ngbrgs7IzsCh@0?P7N0iRT`gIJ7^3{F9Dzy)wP=-E6&713m%=NHWczFr%fe!mc<ollqGw+-*Aa9}(5x`mo+yu%uaJ}|O<e}y"
+"r+)e|TsVwW<W`Y$MHqrUS?Bk1i$JltdpQo^e9zdx?PqEgSlLm&DcA1%s|ZE0nBmF=)?vuP=-O#1q1uTP_m>#z<09cp`&3#i8H8|R"
+"ErMXa<eTCJ`&Qa>?UYjucWSmYIgY5lR08_;YmO;KoiN4`Wmn-4g0pg1k?{{Cp*aI=nVKxzIKh{C+G_k0TO~>v{2^BJ<^HJ(84)G^"
+"JgjBbsGKE=zNy?pXbwfMZXWm+VpNpumvq$R6;n5QR-e<mZl(RT<%ZN2_+PB;?Fcyhc%|zpONA<VWjwo1rQr2Z4$JabZWKJm;NnNM"
+"nd-siWEV9op4{xS289aIpdhwRE&H@Q<A!NbC<kdDSu3PxER2>kV$*IqSh31s+mjNe&HH}Sd{d_A%tOhNgCYi^u_!l}@qAgFJcO@`"
+"&>9WG_dh&5SNQPg&@abmkSX+$MW~`B)XqIN$N+rJ2}47{JBHxemV$hgJfHJzOHPoKXJ%w!+qQU_El@RU%q=c^lukmB1C9%r|6zv7"
+"1gVK5z;9OLEruYeI#P*?lslZxdB+^JmCBq1Ri>*RTaFu!UJUvf=(ol{H?{F<u@elcq{&bxM}(x<T!eweGh@IQ4$$vfLr2@Cw}e+o"
+"M`qhIMxumre-Bu&tK+S%!EH!xvu^=7cb@v2EdMOW7b@mRlk5cKecJMt)?5_u*T6aKmuY0kw}6llK}q{F?CYjvLK4%96f$TRuTwfu"
+"<QwZCY#$98hiK}zQ0lSaSveK@o6x4eQqN_1Q9RN0Zm0!OCV^<3<+kvV!jv_AKC8~=2dXZ#v!p;W6M*&9D`?Qy^ml9E05IK>-n4;~"
+"#Us@LrnX0Ip?)S*?!m~md|zA`{l2Q_Ag!IQG6?wE1)%|29cp4Uz;8w9e&<3ztux38aIT#pvofk4z_E0%s)zS3WnhB7@sHagXa*KE"
+"bd)~o7gp-Dg1Dk)!wdwpGrp*riq55J-~A;QOPyL<GzRg|_MtB7QBBS<VHG4#6NN^gTwXPw{<r=5psO?vf@V?E9lm6elfPNR$24Fz"
+"1sOIqFH&|J(~4ui7Ph^9bo9^$@C-F9xH)L7_-LOx2I0x(w9gqrSAz!KFY2&B^tfzxeMTYOgV(gU!(e!K)BP3UCTjgiKf7mr>X%|M"
+"uS_`Cr3^0>c39hb69>4g=dsKWyq3AZV|~4FBC<F2d$#cDD!Y}!ktGQ|gPv=IfsZ6sN}rSftxxU${!*7Dy7$B-^>&0VE9rMh$sEK5"
+"?G7!&gxfo%x^*#u@TS|)Z(~NFFrUUT7nBRe+tWU}R!2K0iA$jGEm74k)g8AxMvG{aw-{5j`1m&<N9Z^Oo<p$O`+MBOKNgoRmv9R^"
+"Z_uODA(ZK>*j)PviAKGa2Y!0C6T`%pRoRDsd>(BuGO(@|j+8i=+%>FX8G)(H%sy9=_6!pQ+uhw%3})M46-hR$y-BQwj)Uru-@TN8"
+"5@8hb&VYuk^C&o;GHZT_wb+^Sj`CfVb@T9vTN))5S!^p)lDRp9Cas#wK*Y<#FT7!(xaf`b_;ICO1l~6f;Fb$)J_9O?{cI5YeBW^i"
+"t<jXg0yF^~l%pX~u$fSu6#>jNR_et?2ceFeQ?BsGB~Fx)K31FggE<70W7>)0G_c{7H(+N}@3r(9<5Ef$kKTiuP05ka<+OeLEj#*_"
+"G`FIuoUz_DZhl>LVV~w;ikX0p9y?~c9O*8mgtQK&x522sQ&RzGxEAaja`SC`8SSy4O<=vT{_sja?rT0!wS%?ggfi@M{}^U_^5rX%"
+"it?4u<HnoeRJdD<*S+<2!81Q+yJyvHi12jJhwdMWk7j`0hg%<6gSW_!7ezA&X(MuGu8sYhx%j8=`H~F7)$spq6N$^!3?4LC!?xi#"
+"+r6S$C7eyM<P|qd!$Z!K`)D|?<AgT7wlKUM1*+irhJKtVXWbbtCqkbJ)bdH$mJ{!va93vFAd8=)9it1Apg`L8S&##?$6kOXnd$DD"
+"$COXb{HM$Br5PjNf?riE*KTf0S-JyS0S!;U-FNS#)^q2_5*emLsG#(vu+A;ntTPCZP5VDV>dv%<&CUELXl;b{Z<w{M0h~EXU($z7"
+"8<4S}hd71h#}&h=n0EnO^AU~}LL<->KCqawIj$)95BixDQZgw5s{7Ns*)1aI@*lExK2C#j18#0L0rsw2H5K|*ag&jF)TNs^n0~?Q"
+"|64xS%XG%WJFE>oqTCGne%S>|Z{T+#1^$q<n>dCSc*WELV=Dy+@`~R$hfr;DJ80>vTeqIpwqL+}4@5QR*2dqf>f7*e{I@}qq1Ow`"
+"0zqbK*%-aTt`DyGj6~j<+R+J>|CM3uL#^wiJ5`=j7wxu!f@F`J3HRIHo<#sz2OC;ObU+E^{qL6i>FV7J=YzIKXAuetPeZ&;ML@)}"
+"By(J1izyJ`FUBLY@6cV|5JK9N@F_KQ8F4ArQ--JBmDZ9F5g?&%$dFVhZ0gWue1&aRdI)5&ciR|*C($2#jn3C$Z^8Q_;iN46;F%xF"
+"De26#srW9{()}`0f`5Z}LN4a-v^e))i#-ReDqC#auUW$Gmsx71AU<YFhNLn3mi*#wlZ+{9q87XFX?9Kd#8HZSM>ogYqhiB@wKZMl"
+"O3^`8j@%6uwH}?fJu#QPOcR7fIXk2;6*k4cUl4bJ=mdse@hbNz;>W04l^bL+bt}=-eP{@Ill$%S^L<2MW;}@l+G!061$eYl4APUB"
+"?k>jkuwhW@ClO8<pJ=1f^qM-c&@xu}YLnFE?HJ6u&$>Wn@Bg1xqZkpMQWE&qv_<So=WhyNegLF!Pt}K9V0&T}TJBOLy{6He18rtR"
+"n=SRXjnRX$c4*s@QBTVw2BpF$d%>(kA<ao8N^vj2<$-rS8WZ?xatji5{94-Wf=k!Das^f663=~zW_%pyI_)p@4#kMYc0aRcz||Aq"
+"Lym~hByfUZ2O05+=z3Xl&8&VW0z(Rf4w2;DWER83T4%KD093K7{zQftWFH_VlRelyFdbRF$!SybZF}csOGZKb_!n4tk&9(rlX6Fv"
+"y3^vTV@6vvC%{a3@M?Oby1pSg#_h@47cS8$;PHo+5T|h4Of9KtvlUfws_pzscY5<aKbAkR=G^x$=@{Jz9ZB(&bNsds%ZZ_Zy<C~f"
+"fq2GmU6?|x<jX(?A2V|;u*B)#0WX(<HXB;~yu9uU(PqTD%>PSw+vpS{GzO$eR^?@1Tl?=uzE>A$WPm4@lXv%t&<NR(B;(UP-T%Qy"
+"oy>Ebuv1aDWm#!`B%?<6#|89!hOn1LZWiB8ke|4xmoWJ!>W8X{VX7f#g=6Q^+|H!Eob@BT7)4*^MK->M64bKvhI(CHi60wkuMqN-"
+"7vxBxw(=(deJ7Q}fQW^sk+|-&U<d`D{)&N#))tsEQ0w2d3J!Z?>aKk^kNt=?kpwoNd5zGPI(_ebcA(IPWY85{t8fLm*<d93miKAo"
+"lJSPw^?nXd|9js+VqaDJl-hggr8;82fH$-Qw@-IsH%`47pA2{JCv+<Oau@Ra+##&_i8R;jC8N&#TB5vfm{*aOlMjr8Rbqdy8_W=R"
+"oXeSsO1WJ8<9>%?pGKC;grJCdpXk&9(|m~I58(<jIKUf5nCP|kzWyQQbM5Wrh3_1;Nx4aObpifCTG2_S^(kGA?m{QqGkewo)9n9P"
+"Q&R^FCwF|4#6rP+t<n%ct*Tm-F@lpcBTV-fXxrSZ<M|I%Kl<7;*16jN6L@x_-WF}6jzERwWwABsuHc<Gs@#?1FWB%=5jm?NJ}HZ&"
+"3Kh(bkcVb-1YbSHvhzV&vw2147xvSps`~{2+hX*!?Um<1)39~&w)i5o)&;@F5NkyiSlCCVpvxrYJ8e>9W5G6|hI_LQ5r2f4E11O%"
+"DhV))yU|nb)#~jRJ7CV?E9a!w4QD<#aEu8(j)rQCtfvJn(L2Jo9d=5psOo)E;_oBD=n;dwe(SB#UF4*~I?ud4w*M8A(Xhc!&18XL"
+"bx_&4plt}aF1Pb+!`p(C7+KgD5K-AN|3&#jm27KWNL!RE$so0yA&bKYT~3oe<d9L$f&!{J7B0$(&>2aW!(2=phgZV79knw}^G)my"
+"z;KG@ZcB{4t;z7Vez9;^7l`2BRbw}7Tbp?AuN@enuN>T~pyz-_5&Sb$YffrqCatp7E_dT9$<oGD-@H(?Hbi5=BG=`L?>-ilV~$U_"
+"9f&AO_*&wKG?##)w%lp<+7#W3VnlI`U{jg-I|(f~lKf*l`27vImr!BSnak2!3Sg#?*a7o5$sq>XwsMHQ?%#k;n&v9mUM&HH)TuNs"
+"Z$KB*9qh%@EIkYQ<b~=0z&cl2kNyynQnoKi$j<|T3H(A{0|$)~?-Nk4Id#XT34(>M9!~4QXKy#1B+og48<{S?Z304YE>a;`(nWp$"
+"r*$u>#4F{x?HdyYsLw>ZQtZ$5_p_S7nM}oYm(&iyAgdQL*v_70k~Ff6*7uxBKpYV&AN6TS%Axh@oQB(vhGUP#l{c=vn<D~;_BwEn"
+"^77H$fhEGKT0fGw9n-P!w(Ac|&LGq~S|N!AtAD>Tyo9<pi5Yy6D%pzoY%ZC!083U^%#6|0$NEwa;ga4cfyr1ZnvTms>|eV@^y7iN"
+"t)WJBfLjL8c|8KFwi~NZ{#OcPaD0X5Vu;hWBKIJ|kLji#rFa&szySu^4<wJIl1z8^8=q2*1QMoBU>mpd6`!|%DwZ=2XHX=OPV83E"
+"niq_C)$M-xKDwhe93GMAW4`V-lyiu^RX^2d+xp<_?Ew&Gk{}}1u8qz1&m49Kw2>UyrjW!2ptswPoun?JQ(qnOmJRXC9SlUJ;u;UP"
+"N>j2l;ci#BqAFeg^k~BBno&TG*7vRS0e+k#i|W%+aCteqW<|`I?FX}rZ7Z+TdcRaWQmBp{kbM?LRD;(It1Nrq`f5n2bZN_mm=2(|"
+"xrgX(_lq;y-W94^9#{X;?-K^5Gzfk@ew?GFZPy9KMo=x-)kPwe!^jr6m-HtOBJhhy&ooyV%KVnVYA;s(QyL@W;R>H|6<X?Y(|jA{"
+"=TAA4<f1MnP1Rl5Y+VRmNOP<E--_$s#+3+mi?0Dd9GPPX12iMGH2s9Zx=3~L_fk+iQ5XSwLk`x<Wx_An@r@=RD3i;~D9jQwQK2dU"
+"Ix6~)oup=b{{yaHvL&C;a$Ka9S~8`W$^lTR+DBy>dvBnZkt_(tqgVD%rvGw^8X{3)V=%JY1&(T%RtqC0Y^jJ%6*^OGcs&`nC9M}Z"
+">iV|nuOzLXplerSuIYcW4xXd*tH1WWd{-B?)$9MAFoH_e?EL{iyixbVO|-gkzP!b5)FqVxFh`r^b5#mAg6PU3tAvP@fmJ@Z^G2}Q"
+"kn_VO9d&+$a`9gV?3^Oe-uv5l^E+}L>j;Xcm9l!lAH(VOFaAMA*9lQw*DIzP?SlYy#8{SaWx#Zr+Xj7GWH}721Zy%65mH*#>M+s9"
+"r-<;!G%-@RPB^K!%;b5;E;B05!=a&o+UIJjqmU_^PDDyq%Hf$yTXrM_#ib!p7xsAjvz6@KpaC0gKV(&T_dQeCK?@OT&Y(JtpzMh+"
+"bM=#VpVBS#P090xS%ql?cl`<F2+{_M-&zd_yu#=H&y~mDmQ*_hGZ@rTgJxK!)36EX?ad!LG7fz)mNEf+BxK?$+;6UjZ}2IN9u#fN"
+"_a@gQzdYvpGh!lD;v>}s^o)~ItHvd=_d-@DZ`y*egN)Ds>y7pcHt33e6E<F5C|9^PNTdY>f3wu-fwNliwQ|v?2(sP|h%b(`nfSEE"
+"kcA4X#khqCh&5gxbmx8aPo&KW8CX}ZW$i>33eL5oYFI!X?v<%pPWuhUc}!jAJ{S8^dIu(+j$}Q*6qFLGLL+f`$u;(7n)x*|13OCR"
+"Qm4Ita84jbx59c}C5LGN%l&`C=xJAIC1lv;M}2EYdM7+b!KW8PB!w0>Y*?LqQE%?@cqE^n0P5QTJk19@t&#?0U<y0kJFaOO>`XfG"
+";2Rv!>2eoH1J*INznnu$L%^?U?vlKqmNpdi_b9GI`qVjllYC)|EA!U|5b?~)S)WPd*x1q=T_A0--EHUATcQ&Y&dH=7Q58xm)W4OA"
+"t?Yp;-|Z9EPNoq7N|~Ex&2xcO>GV||I^z}GNq-S!p@En+;%(y?P0MT+#@jo##R~G1Ay1G{hbB)J7%Xy78bazOx#MIG*7x3MO+lQ)"
+"YC*C<FqOM0WNK#^VgZTk=~=C0+Y?yqV4r_#$goW*wy%_d05F5dC<vhPILqF9Eu!GaEL0;=()*c%L{veB8yeE<Oh;x}8^wH%wl=Ns"
+">wX7EkIrAHy>bUz5kI#{P5Tp;l6W97m@ay#R+)i%YoRLz@dZD2Knd;!DyVs<sR)cNs>w86M#Kts_8FgA=N#(>g?rGOk=52Bto#}-"
+"pf9}M(4tgh+~H1m_pEq|tUDru?)d)kLUtE3w8^+#Q?S&A%iO(;-M8=gGTB0Nzah@1QKSfxbzp=KmH!3IDjtMqtN>pwM-W0Sxh(ph"
+"(Me|~@Z7qwBEdqmp1YW)PH4&L)HHxV)YE2Xh-<Ctg{YJ%K!41vKD6@#wT`LV+gw)HXzx^6Md_z;CT^y#XEc>YE}?x<fi0s3-1e{J"
+">s=ol!NwOXTHLbX$u&#^8rgRv*C>^ca6^D_02I}|mMHiy?HjvjZwy@fMviy}L#5XhcR>D|`{S2lPn3+drB?j7HaDlEgO**1OW&}U"
+"9-sUb1%~0U*u6~_>vjm(_1+hdJlsbMZ@gn};I!FFbwZXtxqF7l<k)i9h2H4iwiTfc$LKW><G`+O^~{e6EF47rq=F-&>Y0%oq+x5w"
+"Y7u(bLi3TaoP4t(^yF$><D8p9u`rqpT#Qn$)4}0d(+tY6elr<^lqm7{Mv!1%_EYHVk+}xfLj58Gwy$YDJ|}}W3`YksE4wQZwVfo%"
+"p%vBGVkM$(Zuze#Kqo>+`61Ovenw996H*Z96)4bj_N`zzyc>0Q8E=Q#op#AK9=dYK4ad;K_x4rH1z<5MW>?#Ao6TF0jvWYVfUL5r"
+"s>i;oHBz{^f`7~OPaimp{H+X=LAUk$Q!*Wz3J=ZLC<RL(uXV36%`LB+i@PS3TxOO`>Ls|sfnn@dxB8{QuR=vlEW3s}iF6#X$c|VG"
+"4DbeFo{iRi!D>&e3~~uCFf5|J8l)l%4;q0N^>o9q(Tw!djX)^nR-NDnqXdME%S$cxt;%mz^K8Bd;!}0l^|33Ip@21JZ@BI|6WU4$"
+"Tjpol3P4;k%D9^k^bZlu1s8pPQxXeUT*I8aCm4-f);%o=e*S*nl&O>pw-;cs1A<|dVwe<(_7rhwO%L%7E=s&jP%)_w(tz{g25$ta"
+"cOIzBl^O#PdO5#+qRFAqoG}blLMJxHDgheiw5K~r@BuDVkIREbCVURH?okW9#sUBRj~QnG9kHbpGGjcsS-n|<#+_=6NDhU|g{;_~"
+"sjhAso|VR`A)o88LrB69tN99#ShY$_dlM&Dz~7$-!@`HsVVq=qM=HE(4YU8p!_EeI@`G3sZEl(=F33aKE869Q3*UbW`xAdZz1DGe"
+"7ZiC-+pU*kkxIe5UDLMry>^T70}QD%!h1lg1V+yyFV+->a>cg`FK3{YiS-U`t@XC<5mb`Ez$_hsA%C}7GqChdFqHSJ3EbPTm!xy&"
+"tB|X!`>^hA+`P?<noi2z{a#1LPGD^ih6Dku1py_$UY%4}cy9D|Z97X;(+nDItu{o##q#qHFJ;piC%Pbk!~RIWHzus+@l{fLL0uep"
+"Ukp<f%A~B>rF!1(c^Ty<o~ZzB6?N!Ay9TS%4Ul@qQ@1DtC;y0nLeG-ZH6Op0cS<3HGMP5%^Pz1FUV~6eJ)vk;+jptse&eQHQue|3"
+"!TNE$Sfo8uN8aDR3M6b)a^A0plaj#Ku;udEwn!p-?LllUlRpZ~3iOhL()7*3{75c;<19=Ey9F|EgH7NoQRsse#CNcsFvs~*!~v;{"
+"OZfL5Z^t>EJD2bYRUVJsElTN0<J639vn_8B!!(X%qWf94+I`2~vAaPb3p@sakL@GrtC@GC*~l0%3g#yJL0|gVxi+t+4|p8vK8EU?"
+"RjxYbXr?_Ag<S`$_vEsCVgcb@XS;Tz?!H8Fw_r5r*_P_oxw9tjN*ipT=XdY$>W<7eHWTm;z94eu!wS`zY|neZR5~=VWe-Ctwr?}#"
+"|3c|laSUy2VClY@VO&&9V@}a!jMySD0DMsIqqGM`^+;UorQfXLarh__^|1ri*AnCGVPiMMam*XDSmS)W6#gxR%5z1d@P#kE(MR~8"
+"5P$oAut_hU@y0aF<8OiOzQPb~4ZYfMV}AU78fKj|X=vfytM?x|*T5uugs%kaJ~Rr+MnBFI?qC*euBm=P@dfvAKAjkj|017wIhqd%"
+"fdwS7XzQd2NN+4t0=@wAqlKcdtp-k%8PZykI>wqG<A%rct>u`19S}eQUiVE7q0qF;5?WZ_WL{p~M?=ajdW3iV`zrasGn}W@i)yt9"
+"j74CZV-kwmMpeEKy#M4H@yQQdihZ9&)AS|IM3tz39vLXEtvTkF@lbqdB>PkO*%(!&yUsaEW`N_epnRNR8DpB^$OB=r(*mw;Eg0E2"
+")hW%1O*7iH!!S#z3VR#7veYdHhX)~r4yIhf3+Hh_!fXdr0AUy4#@IABw^p}=xdUGQ3X%KZ7Uu3$GOSHhXFO;Xx|gf`>)kz5W%iMg"
+"@6=gvPaU3p9P$7DY*0<qbgu6X$hn$mQGr8l_ok{t^c%=XFjRs*Us8}uS9yo+kYqC%^|?wkOY5eA(IsVv)!+G&E=7hX+qjdmo#qo9"
+"51_riDp6U_P$Bg==xhEOn`hrE$?e|+Ib<@h9&4?;1ShAs+a^Ebt(DcIn)y|%vDs^LalQzUYnthcFaxA!n_dhG)!N#J14RwOZ8p?~"
+"*$hkeRLIURoR*%o4BQ-Mn)+8*-$v#tDO$q6I}jzJb_2Mksh)4ObG8UWOHoK<+!AK<IRG`AhLP@xzvlU+5+i{3{pA1>ZXFw_@oRC#"
+"H4|DopOXu)1?dgtayj0{{pj0>M6H~wAM@Bk99TMAV|$yab@bZ2&h_=Oy6g$UrMBBlS$Mt+ER1}|6lt;VHR<&uep;An9WNU48-q`U"
+"DZmjOZERCV5wn9iM_OzeyAS9brw!48WU~s!nMtjPHxw{O;kly8I%P(?Ya&v{gVa>*%s`#@_#7JFGY@6Kf?W%o0&!f}8zT{IL?m%t"
+"r>wICK5nU-@&h)FNd>rn!SonT_IQ|tzOoP}CnU{6cwJKA<~t4!;v~aUbDDx5`A5;-6|K&}TpLZE`gUhq`lFbaQiXbtKru`X9n_3*"
+"wx0lq9B_QJU<oIyf}ox$xUtoMalu@!@ofLJ!sGHF8nKd5ClLV_I?^9Ml7Fy{MDQr%Q^wi;RO|7Nj>MGQoDD1yDuD=Rev+PRl+Ee2"
+"`qiqRwvII8Qb2H39PoXnKX0LN`X8tsC)W~=8#*6U!Rnz&RE;07iE#7b`%cSYAka{hnm<)!j0qXRm#U#>jHoyUA_#~g05AnCHwycb"
+"08W>Q?mg^9W!A(w`>43V4bT!%bq<ftX&x>@5GJ_F8rS8Hr*dO3pBg=6Ws9no{7e<l4a%D{tfXkbH!TJ?Do-}C>0JNdx$J$$sOd2>"
+"?*IT>xWYK_U79#7ZvB<Fs~VHy%<t`OA?ZK|uqN_HJO*{I3ULw<!iN?*Zg~VFNU<pDj9Jq-imGETI*t2hk%>JNX3r#(MAUGPaEKI`"
+"sG4nIlbR5BJ?ey(!XHOrLQfKpc(O`ZC8H%tJA&~0H+Twu()oi^lYr3w7MV9*I*Am4z^7nDIxkl<tTigX;ELr0O~-*`mdK=8*p?wP"
+"JnxOGj~EN;xxf8>I@<@vg39U(xNexsj(qtKhV@2Qad<fXX`e2&y!*lLZRM`L7x_j`o8_F_OmtTA@QdU>c>y?zAUg@Yk|RqN%sCa&"
+"^wJ;aHfG#{XFg|^J?P=@c&xb#xt%tcLW!jL*RP?T7rs&$b3GZl`>OS~-FSYI(kWF)N@wu3lQm0^;i=D^(vDKOIlBo-Y}@~FRTAA("
+"su1j1Irc)DXsrH!^5<WQ=TM}W=hz4&p7Ydt(x*Q8Y=1ygECOE)Nm?p&4(t(A$$n`$h>9n|G6N~235{;-Is)_|o=hCCGtDq0GbbpZ"
+"9C>9Z?w8a5=a>_0N=gC2S7%ZI{_Th@&c{WtCGYJy08l+(2H9}Nn6o;()tF2XC_qUElYGc`7GkS8fep>I%N!b0T_-i*F|F?Q%;sw{"
+"#!FF1qZCPhCvw1}it+A<kH7*~+|$*J8$4&7@ci4jNZ~W2l;hQ_(~I_uX{3M?w8(B3H;&y|9s;l{8qWl4B@nRH8EJA220SM?yNM|o"
+"%+2Im(*d^0Mo9b=P-8+3%lq1K5{$O-T4=s%Z-CxGiHws<5Hy4Q)ZJag1v|r0F+I2{T;pg^N)#|rx0bSy;{_g&m_rnH7!Xt=xDyTj"
+"Srl{z&Esr|^iC;>8YC=-9@tF4EO^L~3$#T3si*JEjh$ihT6sB~gN%!!ThA4btP&_irAO`%p{6v=ix3+%*TbLki6&>oANlmR;m0+#"
+"IjFai)dJJ6pKJTQ!FA$Ry_iT$Ym$X5JZ4pTB&J=0qybQ-<lpD~Df8*o9?m^~HgG;Qiu2>9j_B5HZCVHnX_Puwpf$N$Aqd#M)Kb$q"
+"cSQMaX^{NZA)iW1H#2~Z4q9UE{4w|QQH>Ph7aNFh7G8X<QXk`)2DX_1=3Hou2@^OPJ0S>SY+;*S8;xT#4j(250~T@jsTDU|Pdk;N"
+"a|V`+7SY~g!(JcB(0^g}6R0@P+WluzFY`wsLRgRs+0JZ!`#2>V2&=?5n`IG~imq>%>0a-gA)FKr`VdHxf5WVe@z2~W@B_K!a{$GF"
+"k{)^2F!{VgN(MCVlHV9b8mEFdOqU0cC8Lq_N`GTIv{LfyQHfO!F1KM2E}m)ytaZyh@*D-HzjC?|N;kp8w>7#dEBpNV0Ad?yrQqgC"
+"clzP@j5g=o4^X;!2+cQ@+2HV=e=>BE|Kp0+GPfu8dTsCEgAWXr9j%M%;src(1HZwN7Kk?8P*Q}kO>0ZQ-R=)cL!;h?Ub`7kd*c1u"
+"O8-+eu4jVI`$?V?mFGq>dHF{TfrRp&&KqZlmogphm8$tx-Y(eJ|6G=0f=7XWc<(RN3B^=tiZGk+>ByWVK}<Pf#0opPtTH!iJo<?-"
+"xUxIpqcD8Uq@a|!D}9R%W#o;#p@$?ue8t>N0SwZDj9L8yczTE?Z7Iu+(nq^-I4p9L8_s*Eni(TfxXqMY4T;TaOEJ(UBZy);N+Ta|"
+"t!$nTi3BID4wbz}`&WoD6)jSd_Wo(yUwzQG6`(F6!4IJWZ1#B$V6x1lMH<%-K<1K8WZ=Kf@XHEcP1`1XUF`P9%GdqlYuxLTPSCyy"
+"&ffavGQ1XLxn4F`yreUqeSQ}H0l?NC!qCLz)d6-+3W~5BE4lQnkY>xlr7(j*Nryz1OysyevU-lFCEr6sm@E!xi<ixuM3g!{4_`Ma"
+"z<o1<djX$IJOWiY_{Gq_0tKKj-kW_#Yz~VWnYlFzK1r0663soO3{C(")),format=L.FORMAT_RAW,filters=[{"id":L.FILTER_LZMA2}]))
